Network.AWS.LexModels

Description

Amazon Lex Build-Time Actions

Amazon Lex is an AWS service for building conversational voice and text interfaces. Use these actions to create, update, and delete conversational bots for new and existing client applications.

Service Configuration

lexModels :: Service

API version 2017-04-19 of the Amazon Lex Model Building Service SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by LexModels.

PreconditionFailedException

_PreconditionFailedException :: AsError a => Getting (First ServiceError) a ServiceError

The checksum of the resource that you are trying to change does not match the checksum in the request. Check the resource's checksum and try again.

ConflictException

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

There was a conflict processing the request. Try your request again.

NotFoundException

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The resource specified in the request was not found. Check the resource and try again.

InternalFailureException

_InternalFailureException :: AsError a => Getting (First ServiceError) a ServiceError

An internal Amazon Lex error occurred. Try your request again.

BadRequestException

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request is not well formed. For example, a value is invalid or a required field is missing. Check the field values, and try again.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The request exceeded a limit. Try your request again.

ResourceInUseException

_ResourceInUseException :: AsError a => Getting (First ServiceError) a ServiceError

The resource that you are attempting to delete is referred to by another resource. Use this information to remove references to the resource that you are trying to delete.

The body of the exception contains a JSON object that describes the resource.

{ "resourceType": BOT | BOTALIAS | BOTCHANNEL | INTENT,

"resourceReference": {

"name": string , "version": string } }

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

DeleteIntentVersion

module Network.AWS.LexModels.DeleteIntentVersion

GetBotAliases

module Network.AWS.LexModels.GetBotAliases

DeleteBotChannelAssociation

module Network.AWS.LexModels.DeleteBotChannelAssociation

CreateSlotTypeVersion

module Network.AWS.LexModels.CreateSlotTypeVersion

GetIntent

module Network.AWS.LexModels.GetIntent

PutIntent

module Network.AWS.LexModels.PutIntent

DeleteIntent

module Network.AWS.LexModels.DeleteIntent

GetBuiltinIntents

module Network.AWS.LexModels.GetBuiltinIntents

PutBot

module Network.AWS.LexModels.PutBot

DeleteBot

module Network.AWS.LexModels.DeleteBot

GetIntentVersions

module Network.AWS.LexModels.GetIntentVersions

GetBuiltinIntent

module Network.AWS.LexModels.GetBuiltinIntent

PutBotAlias

module Network.AWS.LexModels.PutBotAlias

GetBotVersions

module Network.AWS.LexModels.GetBotVersions

GetBotChannelAssociations

module Network.AWS.LexModels.GetBotChannelAssociations

DeleteBotAlias

module Network.AWS.LexModels.DeleteBotAlias

GetSlotTypes

module Network.AWS.LexModels.GetSlotTypes

DeleteUtterances

module Network.AWS.LexModels.DeleteUtterances

GetBots

module Network.AWS.LexModels.GetBots

GetBot

module Network.AWS.LexModels.GetBot

CreateBotVersion

module Network.AWS.LexModels.CreateBotVersion

DeleteSlotTypeVersion

module Network.AWS.LexModels.DeleteSlotTypeVersion

DeleteBotVersion

module Network.AWS.LexModels.DeleteBotVersion

GetSlotType

module Network.AWS.LexModels.GetSlotType

CreateIntentVersion

module Network.AWS.LexModels.CreateIntentVersion

DeleteSlotType

module Network.AWS.LexModels.DeleteSlotType

GetBotChannelAssociation

module Network.AWS.LexModels.GetBotChannelAssociation

PutSlotType

module Network.AWS.LexModels.PutSlotType

GetBuiltinSlotTypes

module Network.AWS.LexModels.GetBuiltinSlotTypes

GetUtterancesView

module Network.AWS.LexModels.GetUtterancesView

GetSlotTypeVersions

module Network.AWS.LexModels.GetSlotTypeVersions

GetIntents

module Network.AWS.LexModels.GetIntents

GetBotAlias

module Network.AWS.LexModels.GetBotAlias

Types

ChannelType

data ChannelType

Constructors

Facebook

Slack

TwilioSms

ContentType

data ContentType

Constructors

PlainText

Ssml

FulfillmentActivityType

data FulfillmentActivityType

Constructors

CodeHook

ReturnIntent

LexStatus

data LexStatus

Constructors

Building

Failed

NotBuilt

Ready

Locale

data Locale

Constructors

EnUs

ProcessBehavior

data ProcessBehavior

Constructors

Build

Save

SlotConstraint

data SlotConstraint

Constructors

Optional

Required

StatusType

data StatusType

Constructors

Detected

Missed

BotAliasMetadata

data BotAliasMetadata

Provides information about a bot alias.

See: botAliasMetadata smart constructor.

botAliasMetadata :: BotAliasMetadata

Creates a value of BotAliasMetadata with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bamChecksum - Checksum of the bot alias.

bamBotVersion - The version of the Amazon Lex bot to which the alias points.

bamBotName - The name of the bot to which the alias points.

bamCreatedDate - The date that the bot alias was created.

bamName - The name of the bot alias.

bamLastUpdatedDate - The date that the bot alias was updated. When you create a resource, the creation date and last updated date are the same.

bamDescription - A description of the bot alias.

bamChecksum :: Lens' BotAliasMetadata (Maybe Text)

Checksum of the bot alias.

bamBotVersion :: Lens' BotAliasMetadata (Maybe Text)

The version of the Amazon Lex bot to which the alias points.

bamBotName :: Lens' BotAliasMetadata (Maybe Text)

The name of the bot to which the alias points.

bamCreatedDate :: Lens' BotAliasMetadata (Maybe UTCTime)

The date that the bot alias was created.

bamName :: Lens' BotAliasMetadata (Maybe Text)

The name of the bot alias.

bamLastUpdatedDate :: Lens' BotAliasMetadata (Maybe UTCTime)

The date that the bot alias was updated. When you create a resource, the creation date and last updated date are the same.

bamDescription :: Lens' BotAliasMetadata (Maybe Text)

A description of the bot alias.

BotChannelAssociation

data BotChannelAssociation

Represents an association between an Amazon Lex bot and an external messaging platform.

See: botChannelAssociation smart constructor.

botChannelAssociation :: BotChannelAssociation

Creates a value of BotChannelAssociation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bcaBotAlias - An alias pointing to the specific version of the Amazon Lex bot to which this association is being made.

bcaBotName - The name of the Amazon Lex bot to which this association is being made.

bcaBotConfiguration - Provides information necessary to communicate with the messaging platform.

bcaCreatedDate - The date that the association between the Amazon Lex bot and the channel was created.

bcaName - The name of the association between the bot and the channel.

bcaType - Specifies the type of association by indicating the type of channel being established between the Amazon Lex bot and the external messaging platform.

bcaDescription - A text description of the association you are creating.

bcaBotAlias :: Lens' BotChannelAssociation (Maybe Text)

An alias pointing to the specific version of the Amazon Lex bot to which this association is being made.

bcaBotName :: Lens' BotChannelAssociation (Maybe Text)

The name of the Amazon Lex bot to which this association is being made.

bcaBotConfiguration :: Lens' BotChannelAssociation (HashMap Text Text)

Provides information necessary to communicate with the messaging platform.

bcaCreatedDate :: Lens' BotChannelAssociation (Maybe UTCTime)

The date that the association between the Amazon Lex bot and the channel was created.

bcaName :: Lens' BotChannelAssociation (Maybe Text)

The name of the association between the bot and the channel.

bcaType :: Lens' BotChannelAssociation (Maybe ChannelType)

Specifies the type of association by indicating the type of channel being established between the Amazon Lex bot and the external messaging platform.

bcaDescription :: Lens' BotChannelAssociation (Maybe Text)

A text description of the association you are creating.

BotMetadata

data BotMetadata

Provides information about a bot. .

See: botMetadata smart constructor.

botMetadata :: BotMetadata

Creates a value of BotMetadata with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bmStatus - The status of the bot.

bmCreatedDate - The date that the bot was created.

bmName - The name of the bot.

bmVersion - The version of the bot. For a new bot, the version is always > LATEST .

bmLastUpdatedDate - The date that the bot was updated. When you create a bot, the creation date and last updated date are the same.

bmDescription - A description of the bot.

bmStatus :: Lens' BotMetadata (Maybe LexStatus)

The status of the bot.

bmCreatedDate :: Lens' BotMetadata (Maybe UTCTime)

The date that the bot was created.

bmName :: Lens' BotMetadata (Maybe Text)

The name of the bot.

bmVersion :: Lens' BotMetadata (Maybe Text)

The version of the bot. For a new bot, the version is always > LATEST .

bmLastUpdatedDate :: Lens' BotMetadata (Maybe UTCTime)

The date that the bot was updated. When you create a bot, the creation date and last updated date are the same.

bmDescription :: Lens' BotMetadata (Maybe Text)

A description of the bot.

BuiltinIntentMetadata

data BuiltinIntentMetadata

Provides metadata for a built-in intent.

See: builtinIntentMetadata smart constructor.

builtinIntentMetadata :: BuiltinIntentMetadata

Creates a value of BuiltinIntentMetadata with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bimSignature - A unique identifier for the built-in intent. To find the signature for an intent, see Standard Built-in Intents in the Alexa Skills Kit .

bimSupportedLocales - A list of identifiers for the locales that the intent supports.

bimSignature :: Lens' BuiltinIntentMetadata (Maybe Text)

A unique identifier for the built-in intent. To find the signature for an intent, see Standard Built-in Intents in the Alexa Skills Kit .

bimSupportedLocales :: Lens' BuiltinIntentMetadata [Locale]

A list of identifiers for the locales that the intent supports.

BuiltinIntentSlot

data BuiltinIntentSlot

Provides information about a slot used in a built-in intent.

See: builtinIntentSlot smart constructor.

builtinIntentSlot :: BuiltinIntentSlot

Creates a value of BuiltinIntentSlot with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bisName - A list of the slots defined for the intent.

bisName :: Lens' BuiltinIntentSlot (Maybe Text)

A list of the slots defined for the intent.

BuiltinSlotTypeMetadata

data BuiltinSlotTypeMetadata

Provides information about a built in slot type.

See: builtinSlotTypeMetadata smart constructor.

builtinSlotTypeMetadata :: BuiltinSlotTypeMetadata

Creates a value of BuiltinSlotTypeMetadata with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bstmSignature - A unique identifier for the built-in slot type. To find the signature for a slot type, see Slot Type Reference in the Alexa Skills Kit .

bstmSupportedLocales - A list of target locales for the slot.

bstmSignature :: Lens' BuiltinSlotTypeMetadata (Maybe Text)

A unique identifier for the built-in slot type. To find the signature for a slot type, see Slot Type Reference in the Alexa Skills Kit .

bstmSupportedLocales :: Lens' BuiltinSlotTypeMetadata [Locale]

A list of target locales for the slot.

CodeHook

data CodeHook

Specifies a Lambda function that verifies requests to a bot or fulfills the user's request to a bot..

See: codeHook smart constructor.

codeHook

Arguments

:: Text
chUri

-> Text
chMessageVersion

-> CodeHook

Creates a value of CodeHook with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

chUri - The Amazon Resource Name (ARN) of the Lambda function.

chMessageVersion - The version of the request-response that you want Amazon Lex to use to invoke your Lambda function. For more information, see 'using-lambda' .

chUri :: Lens' CodeHook Text

The Amazon Resource Name (ARN) of the Lambda function.

chMessageVersion :: Lens' CodeHook Text

The version of the request-response that you want Amazon Lex to use to invoke your Lambda function. For more information, see 'using-lambda' .

EnumerationValue

data EnumerationValue

Each slot type can have a set of values. Each enumeration value represents a value the slot type can take.

For example, a pizza ordering bot could have a slot type that specifies the type of crust that the pizza should have. The slot type could include the values

thick

thin

stuffed

See: enumerationValue smart constructor.

enumerationValue

Arguments

:: Text
evValue

-> EnumerationValue

Creates a value of EnumerationValue with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

evValue - The value of the slot type.

evValue :: Lens' EnumerationValue Text

The value of the slot type.

FollowUpPrompt

data FollowUpPrompt

A prompt for additional activity after an intent is fulfilled. For example, after the OrderPizza intent is fulfilled, you might prompt the user to find out whether the user wants to order drinks.

See: followUpPrompt smart constructor.

followUpPrompt

Arguments

:: Prompt
fupPrompt

-> Statement
fupRejectionStatement

-> FollowUpPrompt

Creates a value of FollowUpPrompt with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fupPrompt - Prompts for information from the user.

fupRejectionStatement - If the user answers "no" to the question defined in the prompt field, Amazon Lex responds with this statement to acknowledge that the intent was canceled.

fupPrompt :: Lens' FollowUpPrompt Prompt

Prompts for information from the user.

fupRejectionStatement :: Lens' FollowUpPrompt Statement

If the user answers "no" to the question defined in the prompt field, Amazon Lex responds with this statement to acknowledge that the intent was canceled.

FulfillmentActivity

data FulfillmentActivity

Describes how the intent is fulfilled after the user provides all of the information required for the intent. You can provide a Lambda function to process the intent, or you can return the intent information to the client application. We recommend that you use a Lambda function so that the relevant logic lives in the Cloud and limit the client-side code primarily to presentation. If you need to update the logic, you only update the Lambda function; you don't need to upgrade your client application.

Consider the following examples:

In a pizza ordering application, after the user provides all of the information for placing an order, you use a Lambda function to place an order with a pizzeria.

In a gaming application, when a user says "pick up a rock," this information must go back to the client application so that it can perform the operation and update the graphics. In this case, you want Amazon Lex to return the intent data to the client.

See: fulfillmentActivity smart constructor.

fulfillmentActivity

Arguments

:: FulfillmentActivityType
faType

-> FulfillmentActivity

Creates a value of FulfillmentActivity with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

faCodeHook - A description of the Lambda function that is run to fulfill the intent.

faType - How the intent should be fulfilled, either by running a Lambda function or by returning the slot data to the client application.

faCodeHook :: Lens' FulfillmentActivity (Maybe CodeHook)

A description of the Lambda function that is run to fulfill the intent.

faType :: Lens' FulfillmentActivity FulfillmentActivityType

How the intent should be fulfilled, either by running a Lambda function or by returning the slot data to the client application.

Intent

data Intent

Identifies the specific version of an intent.

See: intent smart constructor.

intent

Arguments

:: Text
iIntentName

-> Text
iIntentVersion

-> Intent

Creates a value of Intent with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

iIntentName - The name of the intent.

iIntentVersion - The version of the intent.

iIntentName :: Lens' Intent Text

The name of the intent.

iIntentVersion :: Lens' Intent Text

The version of the intent.

IntentMetadata

data IntentMetadata

Provides information about an intent.

See: intentMetadata smart constructor.

intentMetadata :: IntentMetadata

Creates a value of IntentMetadata with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

imCreatedDate - The date that the intent was created.

imName - The name of the intent.

imVersion - The version of the intent.

imLastUpdatedDate - The date that the intent was updated. When you create an intent, the creation date and last updated date are the same.

imDescription - A description of the intent.

imCreatedDate :: Lens' IntentMetadata (Maybe UTCTime)

The date that the intent was created.

imName :: Lens' IntentMetadata (Maybe Text)

The name of the intent.

imVersion :: Lens' IntentMetadata (Maybe Text)

The version of the intent.

imLastUpdatedDate :: Lens' IntentMetadata (Maybe UTCTime)

The date that the intent was updated. When you create an intent, the creation date and last updated date are the same.

imDescription :: Lens' IntentMetadata (Maybe Text)

A description of the intent.

Message

data Message

The message object that provides the message text and its type.

See: message smart constructor.

message

Arguments

:: ContentType
mContentType

-> Text
mContent

-> Message

Creates a value of Message with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mContentType - The content type of the message string.

mContent - The text of the message.

mContentType :: Lens' Message ContentType

The content type of the message string.

mContent :: Lens' Message Text

The text of the message.

Prompt

data Prompt

Obtains information from the user. To define a prompt, provide one or more messages and specify the number of attempts to get information from the user. If you provide more than one message, Amazon Lex chooses one of the messages to use to prompt the user. For more information, see 'how-it-works' .

See: prompt smart constructor.

prompt

Arguments

:: NonEmpty Message
pMessages

-> Natural
pMaxAttempts

-> Prompt

Creates a value of Prompt with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pResponseCard - A response card. Amazon Lex uses this prompt at runtime, in the PostText API response. It substitutes session attributes and slot values for placeholders in the response card. For more information, see 'ex-resp-card' .

pMessages - An array of objects, each of which provides a message string and its type. You can specify the message string in plain text or in Speech Synthesis Markup Language (SSML).

pMaxAttempts - The number of times to prompt the user for information.

pResponseCard :: Lens' Prompt (Maybe Text)

A response card. Amazon Lex uses this prompt at runtime, in the PostText API response. It substitutes session attributes and slot values for placeholders in the response card. For more information, see 'ex-resp-card' .

pMessages :: Lens' Prompt (NonEmpty Message)

An array of objects, each of which provides a message string and its type. You can specify the message string in plain text or in Speech Synthesis Markup Language (SSML).

pMaxAttempts :: Lens' Prompt Natural

The number of times to prompt the user for information.

Slot

data Slot

Identifies the version of a specific slot.

See: slot smart constructor.

slot

Arguments

:: Text
sName

-> SlotConstraint
sSlotConstraint

-> Slot

Creates a value of Slot with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sSlotType - The type of the slot, either a custom slot type that you defined or one of the built-in slot types.

sValueElicitationPrompt - The prompt that Amazon Lex uses to elicit the slot value from the user.

sResponseCard - A set of possible responses for the slot type used by text-based clients. A user chooses an option from the response card, instead of using text to reply.

sPriority - Directs Lex the order in which to elicit this slot value from the user. For example, if the intent has two slots with priorities 1 and 2, AWS Lex first elicits a value for the slot with priority 1. If multiple slots share the same priority, the order in which Lex elicits values is arbitrary.

sSlotTypeVersion - The version of the slot type.

sSampleUtterances - If you know a specific pattern with which users might respond to an Amazon Lex request for a slot value, you can provide those utterances to improve accuracy. This is optional. In most cases, Amazon Lex is capable of understanding user utterances.

sDescription - A description of the slot.

sName - The name of the slot.

sSlotConstraint - Specifies whether the slot is required or optional.

sSlotType :: Lens' Slot (Maybe Text)

The type of the slot, either a custom slot type that you defined or one of the built-in slot types.

sValueElicitationPrompt :: Lens' Slot (Maybe Prompt)

The prompt that Amazon Lex uses to elicit the slot value from the user.

sResponseCard :: Lens' Slot (Maybe Text)

A set of possible responses for the slot type used by text-based clients. A user chooses an option from the response card, instead of using text to reply.

sPriority :: Lens' Slot (Maybe Natural)

Directs Lex the order in which to elicit this slot value from the user. For example, if the intent has two slots with priorities 1 and 2, AWS Lex first elicits a value for the slot with priority 1. If multiple slots share the same priority, the order in which Lex elicits values is arbitrary.

sSlotTypeVersion :: Lens' Slot (Maybe Text)

The version of the slot type.

sSampleUtterances :: Lens' Slot [Text]

If you know a specific pattern with which users might respond to an Amazon Lex request for a slot value, you can provide those utterances to improve accuracy. This is optional. In most cases, Amazon Lex is capable of understanding user utterances.

sDescription :: Lens' Slot (Maybe Text)

A description of the slot.

sName :: Lens' Slot Text

The name of the slot.

sSlotConstraint :: Lens' Slot SlotConstraint

Specifies whether the slot is required or optional.

SlotTypeMetadata

data SlotTypeMetadata

Provides information about a slot type..

See: slotTypeMetadata smart constructor.

slotTypeMetadata :: SlotTypeMetadata

Creates a value of SlotTypeMetadata with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

stmCreatedDate - The date that the slot type was created.

stmName - The name of the slot type.

stmVersion - The version of the slot type.

stmLastUpdatedDate - The date that the slot type was updated. When you create a resource, the creation date and last updated date are the same.

stmDescription - A description of the slot type.

stmCreatedDate :: Lens' SlotTypeMetadata (Maybe UTCTime)

The date that the slot type was created.

stmName :: Lens' SlotTypeMetadata (Maybe Text)

The name of the slot type.

stmVersion :: Lens' SlotTypeMetadata (Maybe Text)

The version of the slot type.

stmLastUpdatedDate :: Lens' SlotTypeMetadata (Maybe UTCTime)

The date that the slot type was updated. When you create a resource, the creation date and last updated date are the same.

stmDescription :: Lens' SlotTypeMetadata (Maybe Text)

A description of the slot type.

Statement

data Statement

A collection of messages that convey information to the user. At runtime, Amazon Lex selects the message to convey.

See: statement smart constructor.

statement

Arguments

:: NonEmpty Message
staMessages

-> Statement

Creates a value of Statement with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

staResponseCard - At runtime, if the client is using the PostText API, Amazon Lex includes the response card in the response. It substitutes all of the session attributes and slot values for placeholders in the response card.

staMessages - A collection of message objects.

staResponseCard :: Lens' Statement (Maybe Text)

At runtime, if the client is using the PostText API, Amazon Lex includes the response card in the response. It substitutes all of the session attributes and slot values for placeholders in the response card.

staMessages :: Lens' Statement (NonEmpty Message)

A collection of message objects.

UtteranceData

data UtteranceData

Provides information about a single utterance that was made to your bot.

See: utteranceData smart constructor.

utteranceData :: UtteranceData

Creates a value of UtteranceData with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

udFirstUtteredDate - The date that the utterance was first recorded.

udCount - The number of times that the utterance was processed.

udUtteranceString - The text that was entered by the user or the text representation of an audio clip.

udLastUtteredDate - The date that the utterance was last recorded.

udDistinctUsers - The total number of individuals that used the utterance.

udFirstUtteredDate :: Lens' UtteranceData (Maybe UTCTime)

The date that the utterance was first recorded.

udCount :: Lens' UtteranceData (Maybe Int)

The number of times that the utterance was processed.

udUtteranceString :: Lens' UtteranceData (Maybe Text)

The text that was entered by the user or the text representation of an audio clip.

udLastUtteredDate :: Lens' UtteranceData (Maybe UTCTime)

The date that the utterance was last recorded.

udDistinctUsers :: Lens' UtteranceData (Maybe Int)

The total number of individuals that used the utterance.

UtteranceList

data UtteranceList

Provides a list of utterances that have been made to a specific version of your bot. The list contains a maximum of 100 utterances.

See: utteranceList smart constructor.

utteranceList :: UtteranceList

Creates a value of UtteranceList with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ulBotVersion - The version of the bot that processed the list.

ulUtterances - One or more UtteranceData objects that contain information about the utterances that have been made to a bot. The maximum number of object is 100.

ulBotVersion :: Lens' UtteranceList (Maybe Text)

The version of the bot that processed the list.

ulUtterances :: Lens' UtteranceList [UtteranceData]

One or more UtteranceData objects that contain information about the utterances that have been made to a bot. The maximum number of object is 100.