Network.AWS.LexModels.CreateBotVersion

Description

Creates a new version of the bot based on the > LATEST version. If the > LATEST version of this resource hasn't changed since you created the last version, Amazon Lex doesn't create a new version. It returns the last created version.

When you create the first version of a bot, Amazon Lex sets the version to 1. Subsequent versions increment by 1. For more information, see 'versioning-intro' .

This operation requires permission for the lex:CreateBotVersion action.

Creating a Request

createBotVersion

Arguments

:: Text
cbvName

-> CreateBotVersion

Creates a value of CreateBotVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cbvChecksum - Identifies a specific revision of the > LATEST version of the bot. If you specify a checksum and the > LATEST version of the bot has a different checksum, a PreconditionFailedException exception is returned and Amazon Lex doesn't publish a new version. If you don't specify a checksum, Amazon Lex publishes the > LATEST version.

cbvName - The name of the bot that you want to create a new version of. The name is case sensitive.

data CreateBotVersion

See: createBotVersion smart constructor.

Request Lenses

cbvChecksum :: Lens' CreateBotVersion (Maybe Text)

Identifies a specific revision of the > LATEST version of the bot. If you specify a checksum and the > LATEST version of the bot has a different checksum, a PreconditionFailedException exception is returned and Amazon Lex doesn't publish a new version. If you don't specify a checksum, Amazon Lex publishes the > LATEST version.

cbvName :: Lens' CreateBotVersion Text

The name of the bot that you want to create a new version of. The name is case sensitive.

Destructuring the Response

createBotVersionResponse

Arguments

:: Int
cbvrsResponseStatus

-> CreateBotVersionResponse

Creates a value of CreateBotVersionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cbvrsFailureReason - If status is FAILED , Amazon Lex provides the reason that it failed to build the bot.

cbvrsStatus - When you send a request to create or update a bot, Amazon Lex sets the status response element to BUILDING . After Amazon Lex builds the bot, it sets status to READY . If Amazon Lex can't build the bot, it sets status to FAILED . Amazon Lex returns the reason for the failure in the failureReason response element.

cbvrsAbortStatement - The message that Amazon Lex uses to abort a conversation. For more information, see PutBot .

cbvrsIntents - An array of Intent objects. For more information, see PutBot .

cbvrsChecksum - Checksum identifying the version of the bot that was created.

cbvrsLocale - Specifies the target locale for the bot.

cbvrsCreatedDate - The date when the bot version was created.

cbvrsName - The name of the bot.

cbvrsVersion - The version of the bot.

cbvrsIdleSessionTTLInSeconds - The maximum time in seconds that Amazon Lex retains the data gathered in a conversation. For more information, see PutBot .

cbvrsClarificationPrompt - The message that Amazon Lex uses when it doesn't understand the user's request. For more information, see PutBot .

cbvrsVoiceId - The Amazon Polly voice ID that Amazon Lex uses for voice interactions with the user.

cbvrsLastUpdatedDate - The date when the > LATEST version of this bot was updated.

cbvrsChildDirected - For each Amazon Lex bot created with the Amazon Lex Model Building Service, you must specify whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to the Children's Online Privacy Protection Act (COPPA) by specifying true or false in the childDirected field. By specifying true in the childDirected field, you confirm that your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. By specifying false in the childDirected field, you confirm that your use of Amazon Lex is not related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. You may not specify a default value for the childDirected field that does not accurately reflect whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. If your use of Amazon Lex relates to a website, program, or other application that is directed in whole or in part, to children under age 13, you must obtain any required verifiable parental consent under COPPA. For information regarding the use of Amazon Lex in connection with websites, programs, or other applications that are directed or targeted, in whole or in part, to children under age 13, see the Amazon Lex FAQ.

cbvrsDescription - A description of the bot.

cbvrsResponseStatus - -- | The response status code.

data CreateBotVersionResponse

See: createBotVersionResponse smart constructor.

Response Lenses

cbvrsFailureReason :: Lens' CreateBotVersionResponse (Maybe Text)

If status is FAILED , Amazon Lex provides the reason that it failed to build the bot.

cbvrsStatus :: Lens' CreateBotVersionResponse (Maybe LexStatus)

When you send a request to create or update a bot, Amazon Lex sets the status response element to BUILDING . After Amazon Lex builds the bot, it sets status to READY . If Amazon Lex can't build the bot, it sets status to FAILED . Amazon Lex returns the reason for the failure in the failureReason response element.

cbvrsAbortStatement :: Lens' CreateBotVersionResponse (Maybe Statement)

The message that Amazon Lex uses to abort a conversation. For more information, see PutBot .

cbvrsIntents :: Lens' CreateBotVersionResponse (Maybe (NonEmpty Intent))

An array of Intent objects. For more information, see PutBot .

cbvrsChecksum :: Lens' CreateBotVersionResponse (Maybe Text)

Checksum identifying the version of the bot that was created.

cbvrsLocale :: Lens' CreateBotVersionResponse (Maybe Locale)

Specifies the target locale for the bot.

cbvrsCreatedDate :: Lens' CreateBotVersionResponse (Maybe UTCTime)

The date when the bot version was created.

cbvrsName :: Lens' CreateBotVersionResponse (Maybe Text)

The name of the bot.

cbvrsVersion :: Lens' CreateBotVersionResponse (Maybe Text)

The version of the bot.

cbvrsIdleSessionTTLInSeconds :: Lens' CreateBotVersionResponse (Maybe Natural)

The maximum time in seconds that Amazon Lex retains the data gathered in a conversation. For more information, see PutBot .

cbvrsClarificationPrompt :: Lens' CreateBotVersionResponse (Maybe Prompt)

The message that Amazon Lex uses when it doesn't understand the user's request. For more information, see PutBot .

cbvrsVoiceId :: Lens' CreateBotVersionResponse (Maybe Text)

The Amazon Polly voice ID that Amazon Lex uses for voice interactions with the user.

cbvrsLastUpdatedDate :: Lens' CreateBotVersionResponse (Maybe UTCTime)

The date when the > LATEST version of this bot was updated.

cbvrsChildDirected :: Lens' CreateBotVersionResponse (Maybe Bool)

For each Amazon Lex bot created with the Amazon Lex Model Building Service, you must specify whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to the Children's Online Privacy Protection Act (COPPA) by specifying true or false in the childDirected field. By specifying true in the childDirected field, you confirm that your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. By specifying false in the childDirected field, you confirm that your use of Amazon Lex is not related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. You may not specify a default value for the childDirected field that does not accurately reflect whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. If your use of Amazon Lex relates to a website, program, or other application that is directed in whole or in part, to children under age 13, you must obtain any required verifiable parental consent under COPPA. For information regarding the use of Amazon Lex in connection with websites, programs, or other applications that are directed or targeted, in whole or in part, to children under age 13, see the Amazon Lex FAQ.

cbvrsDescription :: Lens' CreateBotVersionResponse (Maybe Text)

A description of the bot.

cbvrsResponseStatus :: Lens' CreateBotVersionResponse Int

- | The response status code.