Network.AWS.LexModels.CreateIntentVersion

Description

Creates a new version of an intent based on the > LATEST version of the intent. If the > LATEST version of this intent hasn't changed since you last updated it, Amazon Lex doesn't create a new version. It returns the last version you created.

When you create a version of an intent, Amazon Lex sets the version to 1. Subsequent versions increment by 1. For more information, see 'versioning-intro' .

This operation requires permissions to perform the lex:CreateIntentVersion action.

Creating a Request

createIntentVersion

Arguments

:: Text
civName

-> CreateIntentVersion

Creates a value of CreateIntentVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

civChecksum - Checksum of the > LATEST version of the intent that should be used to create the new version. If you specify a checksum and the > LATEST version of the intent has a different checksum, Amazon Lex returns a PreconditionFailedException exception and doesn't publish a new version. If you don't specify a checksum, Amazon Lex publishes the > LATEST version.

civName - The name of the intent that you want to create a new version of. The name is case sensitive.

data CreateIntentVersion

See: createIntentVersion smart constructor.

Request Lenses

civChecksum :: Lens' CreateIntentVersion (Maybe Text)

Checksum of the > LATEST version of the intent that should be used to create the new version. If you specify a checksum and the > LATEST version of the intent has a different checksum, Amazon Lex returns a PreconditionFailedException exception and doesn't publish a new version. If you don't specify a checksum, Amazon Lex publishes the > LATEST version.

civName :: Lens' CreateIntentVersion Text

The name of the intent that you want to create a new version of. The name is case sensitive.

Destructuring the Response

createIntentVersionResponse

Arguments

:: Int
civrsResponseStatus

-> CreateIntentVersionResponse

Creates a value of CreateIntentVersionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

civrsFulfillmentActivity - Describes how the intent is fulfilled.

civrsSlots - An array of slot types that defines the information required to fulfill the intent.

civrsRejectionStatement - If the user answers "no" to the question defined in confirmationPrompt , Amazon Lex responds with this statement to acknowledge that the intent was canceled.

civrsChecksum - Checksum of the intent version created.

civrsConclusionStatement - After the Lambda function specified in the fulfillmentActivity field fulfills the intent, Amazon Lex conveys this statement to the user.

civrsSampleUtterances - An array of sample utterances configured for the intent.

civrsParentIntentSignature - A unique identifier for a built-in intent.

civrsCreatedDate - The date that the intent was created.

civrsName - The name of the intent.

civrsVersion - The version number assigned to the new version of the intent.

civrsFollowUpPrompt - If defined, Amazon Lex uses this prompt to solicit additional user activity after the intent is fulfilled.

civrsLastUpdatedDate - The date that the intent was updated.

civrsConfirmationPrompt - If defined, the prompt that Amazon Lex uses to confirm the user's intent before fulfilling it.

civrsDialogCodeHook - If defined, Amazon Lex invokes this Lambda function for each user input.

civrsDescription - A description of the intent.

civrsResponseStatus - -- | The response status code.

data CreateIntentVersionResponse

See: createIntentVersionResponse smart constructor.

Response Lenses

civrsFulfillmentActivity :: Lens' CreateIntentVersionResponse (Maybe FulfillmentActivity)

Describes how the intent is fulfilled.

civrsSlots :: Lens' CreateIntentVersionResponse [Slot]

An array of slot types that defines the information required to fulfill the intent.

civrsRejectionStatement :: Lens' CreateIntentVersionResponse (Maybe Statement)

If the user answers "no" to the question defined in confirmationPrompt , Amazon Lex responds with this statement to acknowledge that the intent was canceled.

civrsChecksum :: Lens' CreateIntentVersionResponse (Maybe Text)

Checksum of the intent version created.

civrsConclusionStatement :: Lens' CreateIntentVersionResponse (Maybe Statement)

After the Lambda function specified in the fulfillmentActivity field fulfills the intent, Amazon Lex conveys this statement to the user.

civrsSampleUtterances :: Lens' CreateIntentVersionResponse [Text]

An array of sample utterances configured for the intent.

civrsParentIntentSignature :: Lens' CreateIntentVersionResponse (Maybe Text)

A unique identifier for a built-in intent.

civrsCreatedDate :: Lens' CreateIntentVersionResponse (Maybe UTCTime)

The date that the intent was created.

civrsName :: Lens' CreateIntentVersionResponse (Maybe Text)

The name of the intent.

civrsVersion :: Lens' CreateIntentVersionResponse (Maybe Text)

The version number assigned to the new version of the intent.

civrsFollowUpPrompt :: Lens' CreateIntentVersionResponse (Maybe FollowUpPrompt)

If defined, Amazon Lex uses this prompt to solicit additional user activity after the intent is fulfilled.

civrsLastUpdatedDate :: Lens' CreateIntentVersionResponse (Maybe UTCTime)

The date that the intent was updated.

civrsConfirmationPrompt :: Lens' CreateIntentVersionResponse (Maybe Prompt)

If defined, the prompt that Amazon Lex uses to confirm the user's intent before fulfilling it.

civrsDialogCodeHook :: Lens' CreateIntentVersionResponse (Maybe CodeHook)

If defined, Amazon Lex invokes this Lambda function for each user input.

civrsDescription :: Lens' CreateIntentVersionResponse (Maybe Text)

A description of the intent.

civrsResponseStatus :: Lens' CreateIntentVersionResponse Int

- | The response status code.