Network.AWS.LexModels.CreateSlotTypeVersion

Description

Creates a new version of a slot type based on the > LATEST version of the specified slot type. If the > LATEST version of this resource has not changed since the last version that you created, Amazon Lex doesn't create a new version. It returns the last version that you created.

When you create a version of a slot type, Amazon Lex sets the version to 1. Subsequent versions increment by 1. For more information, see 'versioning-intro' .

This operation requires permissions for the lex:CreateSlotTypeVersion action.

Creating a Request

createSlotTypeVersion

Arguments

:: Text
cstvName

-> CreateSlotTypeVersion

Creates a value of CreateSlotTypeVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cstvChecksum - Checksum for the > LATEST version of the slot type that you want to publish. If you specify a checksum and the > LATEST version of the slot type has a different checksum, Amazon Lex returns a PreconditionFailedException exception and doesn't publish the new version. If you don't specify a checksum, Amazon Lex publishes the > LATEST version.

cstvName - The name of the slot type that you want to create a new version for. The name is case sensitive.

data CreateSlotTypeVersion

See: createSlotTypeVersion smart constructor.

Request Lenses

cstvChecksum :: Lens' CreateSlotTypeVersion (Maybe Text)

Checksum for the > LATEST version of the slot type that you want to publish. If you specify a checksum and the > LATEST version of the slot type has a different checksum, Amazon Lex returns a PreconditionFailedException exception and doesn't publish the new version. If you don't specify a checksum, Amazon Lex publishes the > LATEST version.

cstvName :: Lens' CreateSlotTypeVersion Text

The name of the slot type that you want to create a new version for. The name is case sensitive.

Destructuring the Response

createSlotTypeVersionResponse

Arguments

:: Int
cstvrsResponseStatus

-> CreateSlotTypeVersionResponse

Creates a value of CreateSlotTypeVersionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cstvrsChecksum - Checksum of the > LATEST version of the slot type.

cstvrsCreatedDate - The date that the slot type was created.

cstvrsName - The name of the slot type.

cstvrsVersion - The version assigned to the new slot type version.

cstvrsLastUpdatedDate - The date that the slot type was updated. When you create a resource, the creation date and last update date are the same.

cstvrsDescription - A description of the slot type.

cstvrsEnumerationValues - A list of EnumerationValue objects that defines the values that the slot type can take.

cstvrsResponseStatus - -- | The response status code.

data CreateSlotTypeVersionResponse

See: createSlotTypeVersionResponse smart constructor.

Response Lenses

cstvrsChecksum :: Lens' CreateSlotTypeVersionResponse (Maybe Text)

Checksum of the > LATEST version of the slot type.

cstvrsCreatedDate :: Lens' CreateSlotTypeVersionResponse (Maybe UTCTime)

The date that the slot type was created.

cstvrsName :: Lens' CreateSlotTypeVersionResponse (Maybe Text)

The name of the slot type.

cstvrsVersion :: Lens' CreateSlotTypeVersionResponse (Maybe Text)

The version assigned to the new slot type version.

cstvrsLastUpdatedDate :: Lens' CreateSlotTypeVersionResponse (Maybe UTCTime)

The date that the slot type was updated. When you create a resource, the creation date and last update date are the same.

cstvrsDescription :: Lens' CreateSlotTypeVersionResponse (Maybe Text)

A description of the slot type.

cstvrsEnumerationValues :: Lens' CreateSlotTypeVersionResponse (Maybe (NonEmpty EnumerationValue))

A list of EnumerationValue objects that defines the values that the slot type can take.

cstvrsResponseStatus :: Lens' CreateSlotTypeVersionResponse Int

- | The response status code.