Network.AWS.LexModels.DeleteBotAlias

Description

Deletes an alias for the specified bot.

You can't delete an alias that is used in the association between a bot and a messaging channel. If an alias is used in a channel association, the DeleteBot operation returns a ResourceInUseException exception that includes a reference to the channel association that refers to the bot. You can remove the reference to the alias by deleting the channel association. If you get the same exception again, delete the referring association until the DeleteBotAlias operation is successful.

Creating a Request

deleteBotAlias

Arguments

:: Text
dbaName

-> Text
dbaBotName

-> DeleteBotAlias

Creates a value of DeleteBotAlias with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dbaName - The name of the alias to delete. The name is case sensitive.

dbaBotName - The name of the bot that the alias points to.

data DeleteBotAlias

See: deleteBotAlias smart constructor.

Request Lenses

dbaName :: Lens' DeleteBotAlias Text

The name of the alias to delete. The name is case sensitive.

dbaBotName :: Lens' DeleteBotAlias Text

The name of the bot that the alias points to.

Destructuring the Response

deleteBotAliasResponse :: DeleteBotAliasResponse

Creates a value of DeleteBotAliasResponse with the minimum fields required to make a request.

data DeleteBotAliasResponse

See: deleteBotAliasResponse smart constructor.