Network.AWS.LexModels.DeleteBotChannelAssociation

Description

Deletes the association between an Amazon Lex bot and a messaging platform.

This operation requires permission for the lex:DeleteBotChannelAssociation action.

Creating a Request

deleteBotChannelAssociation

Arguments

:: Text
dbcaName

-> Text
dbcaBotName

-> Text
dbcaBotAlias

-> DeleteBotChannelAssociation

Creates a value of DeleteBotChannelAssociation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dbcaName - The name of the association. The name is case sensitive.

dbcaBotName - The name of the Amazon Lex bot.

dbcaBotAlias - An alias that points to the specific version of the Amazon Lex bot to which this association is being made.

data DeleteBotChannelAssociation

See: deleteBotChannelAssociation smart constructor.

Request Lenses

dbcaName :: Lens' DeleteBotChannelAssociation Text

The name of the association. The name is case sensitive.

dbcaBotName :: Lens' DeleteBotChannelAssociation Text

The name of the Amazon Lex bot.

dbcaBotAlias :: Lens' DeleteBotChannelAssociation Text

An alias that points to the specific version of the Amazon Lex bot to which this association is being made.

Destructuring the Response

deleteBotChannelAssociationResponse :: DeleteBotChannelAssociationResponse

Creates a value of DeleteBotChannelAssociationResponse with the minimum fields required to make a request.

data DeleteBotChannelAssociationResponse

See: deleteBotChannelAssociationResponse smart constructor.