Network.AWS.LexModels.DeleteBotVersion

Description

Deletes a specific version of a bot. To delete all versions of a bot, use the DeleteBot operation.

This operation requires permissions for the lex:DeleteBotVersion action.

Creating a Request

deleteBotVersion

Arguments

:: Text
dbvName

-> Text
dbvVersion

-> DeleteBotVersion

Creates a value of DeleteBotVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dbvName - The name of the bot.

dbvVersion - The version of the bot to delete. You cannot delete the > LATEST version of the bot. To delete the > LATEST version, use the DeleteBot operation.

data DeleteBotVersion

See: deleteBotVersion smart constructor.

Request Lenses

dbvName :: Lens' DeleteBotVersion Text

The name of the bot.

dbvVersion :: Lens' DeleteBotVersion Text

The version of the bot to delete. You cannot delete the > LATEST version of the bot. To delete the > LATEST version, use the DeleteBot operation.

Destructuring the Response

deleteBotVersionResponse :: DeleteBotVersionResponse

Creates a value of DeleteBotVersionResponse with the minimum fields required to make a request.

data DeleteBotVersionResponse

See: deleteBotVersionResponse smart constructor.