Network.AWS.LexModels.DeleteIntentVersion

Description

Deletes a specific version of an intent. To delete all versions of a intent, use the DeleteIntent operation.

This operation requires permissions for the lex:DeleteIntentVersion action.

Creating a Request

deleteIntentVersion

Arguments

:: Text
divName

-> Text
divVersion

-> DeleteIntentVersion

Creates a value of DeleteIntentVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

divName - The name of the intent.

divVersion - The version of the intent to delete. You cannot delete the > LATEST version of the intent. To delete the > LATEST version, use the DeleteIntent operation.

data DeleteIntentVersion

See: deleteIntentVersion smart constructor.

Request Lenses

divName :: Lens' DeleteIntentVersion Text

The name of the intent.

divVersion :: Lens' DeleteIntentVersion Text

The version of the intent to delete. You cannot delete the > LATEST version of the intent. To delete the > LATEST version, use the DeleteIntent operation.

Destructuring the Response

deleteIntentVersionResponse :: DeleteIntentVersionResponse

Creates a value of DeleteIntentVersionResponse with the minimum fields required to make a request.

data DeleteIntentVersionResponse

See: deleteIntentVersionResponse smart constructor.