Network.AWS.LexModels.DeleteSlotTypeVersion

Description

Deletes a specific version of a slot type. To delete all versions of a slot type, use the DeleteSlotType operation.

This operation requires permissions for the lex:DeleteSlotTypeVersion action.

Creating a Request

deleteSlotTypeVersion

Arguments

:: Text
dstvName

-> Text
dstvVersion

-> DeleteSlotTypeVersion

Creates a value of DeleteSlotTypeVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dstvName - The name of the slot type.

dstvVersion - The version of the slot type to delete. You cannot delete the > LATEST version of the slot type. To delete the > LATEST version, use the DeleteSlotType operation.

data DeleteSlotTypeVersion

See: deleteSlotTypeVersion smart constructor.

Request Lenses

dstvName :: Lens' DeleteSlotTypeVersion Text

The name of the slot type.

dstvVersion :: Lens' DeleteSlotTypeVersion Text

The version of the slot type to delete. You cannot delete the > LATEST version of the slot type. To delete the > LATEST version, use the DeleteSlotType operation.

Destructuring the Response

deleteSlotTypeVersionResponse :: DeleteSlotTypeVersionResponse

Creates a value of DeleteSlotTypeVersionResponse with the minimum fields required to make a request.

data DeleteSlotTypeVersionResponse

See: deleteSlotTypeVersionResponse smart constructor.