Network.AWS.LexModels.GetBot

Description

Returns metadata information for a specific bot. You must provide the bot name and the bot version or alias.

The GetBot operation requires permissions for the lex:GetBot action.

Creating a Request

getBot

Arguments

:: Text
gbName

-> Text
gbVersionOrAlias

-> GetBot

Creates a value of GetBot with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbName - The name of the bot. The name is case sensitive.

gbVersionOrAlias - The version or alias of the bot.

data GetBot

See: getBot smart constructor.

Request Lenses

gbName :: Lens' GetBot Text

The name of the bot. The name is case sensitive.

gbVersionOrAlias :: Lens' GetBot Text

The version or alias of the bot.

Destructuring the Response

getBotResponse

Arguments

:: Int
gbrsResponseStatus

-> GetBotResponse

Creates a value of GetBotResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbrsFailureReason - If status is FAILED , Amazon Lex explains why it failed to build the bot.

gbrsStatus - The status of the bot. If the bot is ready to run, the status is READY . If there was a problem with building the bot, the status is FAILED and the failureReason explains why the bot did not build. If the bot was saved but not built, the status is NOT BUILT .

gbrsAbortStatement - The message that Amazon Lex returns when the user elects to end the conversation without completing it. For more information, see PutBot .

gbrsIntents - An array of intent objects. For more information, see PutBot .

gbrsChecksum - Checksum of the bot used to identify a specific revision of the bot's > LATEST version.

gbrsLocale - The target locale for the bot.

gbrsCreatedDate - The date that the bot was created.

gbrsName - The name of the bot.

gbrsVersion - The version of the bot. For a new bot, the version is always > LATEST .

gbrsIdleSessionTTLInSeconds - The maximum time in seconds that Amazon Lex retains the data gathered in a conversation. For more information, see PutBot .

gbrsClarificationPrompt - The message Amazon Lex uses when it doesn't understand the user's request. For more information, see PutBot .

gbrsVoiceId - The Amazon Polly voice ID that Amazon Lex uses for voice interaction with the user. For more information, see PutBot .

gbrsLastUpdatedDate - The date that the bot was updated. When you create a resource, the creation date and last updated date are the same.

gbrsChildDirected - For each Amazon Lex bot created with the Amazon Lex Model Building Service, you must specify whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to the Children's Online Privacy Protection Act (COPPA) by specifying true or false in the childDirected field. By specifying true in the childDirected field, you confirm that your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. By specifying false in the childDirected field, you confirm that your use of Amazon Lex is not related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. You may not specify a default value for the childDirected field that does not accurately reflect whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. If your use of Amazon Lex relates to a website, program, or other application that is directed in whole or in part, to children under age 13, you must obtain any required verifiable parental consent under COPPA. For information regarding the use of Amazon Lex in connection with websites, programs, or other applications that are directed or targeted, in whole or in part, to children under age 13, see the Amazon Lex FAQ.

gbrsDescription - A description of the bot.

gbrsResponseStatus - -- | The response status code.

data GetBotResponse

See: getBotResponse smart constructor.

Response Lenses

gbrsFailureReason :: Lens' GetBotResponse (Maybe Text)

If status is FAILED , Amazon Lex explains why it failed to build the bot.

gbrsStatus :: Lens' GetBotResponse (Maybe LexStatus)

The status of the bot. If the bot is ready to run, the status is READY . If there was a problem with building the bot, the status is FAILED and the failureReason explains why the bot did not build. If the bot was saved but not built, the status is NOT BUILT .

gbrsAbortStatement :: Lens' GetBotResponse (Maybe Statement)

The message that Amazon Lex returns when the user elects to end the conversation without completing it. For more information, see PutBot .

gbrsIntents :: Lens' GetBotResponse (Maybe (NonEmpty Intent))

An array of intent objects. For more information, see PutBot .

gbrsChecksum :: Lens' GetBotResponse (Maybe Text)

Checksum of the bot used to identify a specific revision of the bot's > LATEST version.

gbrsLocale :: Lens' GetBotResponse (Maybe Locale)

The target locale for the bot.

gbrsCreatedDate :: Lens' GetBotResponse (Maybe UTCTime)

The date that the bot was created.

gbrsName :: Lens' GetBotResponse (Maybe Text)

The name of the bot.

gbrsVersion :: Lens' GetBotResponse (Maybe Text)

The version of the bot. For a new bot, the version is always > LATEST .

gbrsIdleSessionTTLInSeconds :: Lens' GetBotResponse (Maybe Natural)

The maximum time in seconds that Amazon Lex retains the data gathered in a conversation. For more information, see PutBot .

gbrsClarificationPrompt :: Lens' GetBotResponse (Maybe Prompt)

The message Amazon Lex uses when it doesn't understand the user's request. For more information, see PutBot .

gbrsVoiceId :: Lens' GetBotResponse (Maybe Text)

The Amazon Polly voice ID that Amazon Lex uses for voice interaction with the user. For more information, see PutBot .

gbrsLastUpdatedDate :: Lens' GetBotResponse (Maybe UTCTime)

The date that the bot was updated. When you create a resource, the creation date and last updated date are the same.

gbrsChildDirected :: Lens' GetBotResponse (Maybe Bool)

For each Amazon Lex bot created with the Amazon Lex Model Building Service, you must specify whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to the Children's Online Privacy Protection Act (COPPA) by specifying true or false in the childDirected field. By specifying true in the childDirected field, you confirm that your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. By specifying false in the childDirected field, you confirm that your use of Amazon Lex is not related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. You may not specify a default value for the childDirected field that does not accurately reflect whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. If your use of Amazon Lex relates to a website, program, or other application that is directed in whole or in part, to children under age 13, you must obtain any required verifiable parental consent under COPPA. For information regarding the use of Amazon Lex in connection with websites, programs, or other applications that are directed or targeted, in whole or in part, to children under age 13, see the Amazon Lex FAQ.

gbrsDescription :: Lens' GetBotResponse (Maybe Text)

A description of the bot.

gbrsResponseStatus :: Lens' GetBotResponse Int

- | The response status code.