Network.AWS.LexModels.GetBotAliases

Description

Returns a list of aliases for a specified Amazon Lex bot.

This operation requires permissions for the lex:GetBotAliases action.

Creating a Request

getBotAliases

Arguments

:: Text
gbaBotName

-> GetBotAliases

Creates a value of GetBotAliases with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbaNameContains - Substring to match in bot alias names. An alias will be returned if any part of its name matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz."

gbaNextToken - A pagination token for fetching the next page of aliases. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of aliases, specify the pagination token in the next request.

gbaMaxResults - The maximum number of aliases to return in the response. The default is 50. .

gbaBotName - The name of the bot.

data GetBotAliases

See: getBotAliases smart constructor.

Request Lenses

gbaNameContains :: Lens' GetBotAliases (Maybe Text)

Substring to match in bot alias names. An alias will be returned if any part of its name matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz."

gbaNextToken :: Lens' GetBotAliases (Maybe Text)

A pagination token for fetching the next page of aliases. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of aliases, specify the pagination token in the next request.

gbaMaxResults :: Lens' GetBotAliases (Maybe Natural)

The maximum number of aliases to return in the response. The default is 50. .

gbaBotName :: Lens' GetBotAliases Text

The name of the bot.

Destructuring the Response

getBotAliasesResponse

Arguments

:: Int
gbarsResponseStatus

-> GetBotAliasesResponse

Creates a value of GetBotAliasesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbarsNextToken - A pagination token for fetching next page of aliases. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of aliases, specify the pagination token in the next request.

gbarsBotAliases - An array of BotAliasMetadata objects, each describing a bot alias.

gbarsResponseStatus - -- | The response status code.

data GetBotAliasesResponse

See: getBotAliasesResponse smart constructor.

Response Lenses

gbarsNextToken :: Lens' GetBotAliasesResponse (Maybe Text)

A pagination token for fetching next page of aliases. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of aliases, specify the pagination token in the next request.

gbarsBotAliases :: Lens' GetBotAliasesResponse [BotAliasMetadata]

An array of BotAliasMetadata objects, each describing a bot alias.

gbarsResponseStatus :: Lens' GetBotAliasesResponse Int

- | The response status code.