Network.AWS.LexModels.GetBotChannelAssociation

Description

Returns information about the association between an Amazon Lex bot and a messaging platform.

This operation requires permissions for the lex:GetBotChannelAssociation action.

Creating a Request

getBotChannelAssociation

Arguments

:: Text
gName

-> Text
gBotName

-> Text
gBotAlias

-> GetBotChannelAssociation

Creates a value of GetBotChannelAssociation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gName - The name of the association between the bot and the channel. The name is case sensitive.

gBotName - The name of the Amazon Lex bot.

gBotAlias - An alias pointing to the specific version of the Amazon Lex bot to which this association is being made.

data GetBotChannelAssociation

See: getBotChannelAssociation smart constructor.

Request Lenses

gName :: Lens' GetBotChannelAssociation Text

The name of the association between the bot and the channel. The name is case sensitive.

gBotName :: Lens' GetBotChannelAssociation Text

The name of the Amazon Lex bot.

gBotAlias :: Lens' GetBotChannelAssociation Text

An alias pointing to the specific version of the Amazon Lex bot to which this association is being made.

Destructuring the Response

getBotChannelAssociationResponse

Arguments

:: Int
gbcarsResponseStatus

-> GetBotChannelAssociationResponse

Creates a value of GetBotChannelAssociationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbcarsBotAlias - An alias pointing to the specific version of the Amazon Lex bot to which this association is being made.

gbcarsBotName - The name of the Amazon Lex bot.

gbcarsBotConfiguration - Provides information that the messaging platform needs to communicate with the Amazon Lex bot.

gbcarsCreatedDate - The date that the association between the bot and the channel was created.

gbcarsName - The name of the association between the bot and the channel.

gbcarsType - The type of the messaging platform.

gbcarsDescription - A description of the association between the bot and the channel.

gbcarsResponseStatus - -- | The response status code.

data GetBotChannelAssociationResponse

See: getBotChannelAssociationResponse smart constructor.

Response Lenses

gbcarsBotAlias :: Lens' GetBotChannelAssociationResponse (Maybe Text)

An alias pointing to the specific version of the Amazon Lex bot to which this association is being made.

gbcarsBotName :: Lens' GetBotChannelAssociationResponse (Maybe Text)

The name of the Amazon Lex bot.

gbcarsBotConfiguration :: Lens' GetBotChannelAssociationResponse (HashMap Text Text)

Provides information that the messaging platform needs to communicate with the Amazon Lex bot.

gbcarsCreatedDate :: Lens' GetBotChannelAssociationResponse (Maybe UTCTime)

The date that the association between the bot and the channel was created.

gbcarsName :: Lens' GetBotChannelAssociationResponse (Maybe Text)

The name of the association between the bot and the channel.

gbcarsType :: Lens' GetBotChannelAssociationResponse (Maybe ChannelType)

The type of the messaging platform.

gbcarsDescription :: Lens' GetBotChannelAssociationResponse (Maybe Text)

A description of the association between the bot and the channel.

gbcarsResponseStatus :: Lens' GetBotChannelAssociationResponse Int

- | The response status code.