Network.AWS.LexModels.GetBotChannelAssociations

Description

Returns a list of all of the channels associated with the specified bot.

The GetBotChannelAssociations operation requires permissions for the lex:GetBotChannelAssociations action.

Creating a Request

getBotChannelAssociations

Arguments

:: Text
gbcaBotName

-> Text
gbcaBotAlias

-> GetBotChannelAssociations

Creates a value of GetBotChannelAssociations with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbcaNameContains - Substring to match in channel association names. An association will be returned if any part of its name matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz." To return all bot channel associations, use a hyphen ("-") as the nameContains parameter.

gbcaNextToken - A pagination token for fetching the next page of associations. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of associations, specify the pagination token in the next request.

gbcaMaxResults - The maximum number of associations to return in the response. The default is 50.

gbcaBotName - The name of the Amazon Lex bot in the association.

gbcaBotAlias - An alias pointing to the specific version of the Amazon Lex bot to which this association is being made.

data GetBotChannelAssociations

See: getBotChannelAssociations smart constructor.

Request Lenses

gbcaNameContains :: Lens' GetBotChannelAssociations (Maybe Text)

Substring to match in channel association names. An association will be returned if any part of its name matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz." To return all bot channel associations, use a hyphen ("-") as the nameContains parameter.

gbcaNextToken :: Lens' GetBotChannelAssociations (Maybe Text)

A pagination token for fetching the next page of associations. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of associations, specify the pagination token in the next request.

gbcaMaxResults :: Lens' GetBotChannelAssociations (Maybe Natural)

The maximum number of associations to return in the response. The default is 50.

gbcaBotName :: Lens' GetBotChannelAssociations Text

The name of the Amazon Lex bot in the association.

gbcaBotAlias :: Lens' GetBotChannelAssociations Text

An alias pointing to the specific version of the Amazon Lex bot to which this association is being made.

Destructuring the Response

getBotChannelAssociationsResponse

Arguments

:: Int
gbcasrsResponseStatus

-> GetBotChannelAssociationsResponse

Creates a value of GetBotChannelAssociationsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbcasrsBotChannelAssociations - An array of objects, one for each association, that provides information about the Amazon Lex bot and its association with the channel.

gbcasrsNextToken - A pagination token that fetches the next page of associations. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of associations, specify the pagination token in the next request.

gbcasrsResponseStatus - -- | The response status code.

data GetBotChannelAssociationsResponse

See: getBotChannelAssociationsResponse smart constructor.

Response Lenses

gbcasrsBotChannelAssociations :: Lens' GetBotChannelAssociationsResponse [BotChannelAssociation]

An array of objects, one for each association, that provides information about the Amazon Lex bot and its association with the channel.

gbcasrsNextToken :: Lens' GetBotChannelAssociationsResponse (Maybe Text)

A pagination token that fetches the next page of associations. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of associations, specify the pagination token in the next request.

gbcasrsResponseStatus :: Lens' GetBotChannelAssociationsResponse Int

- | The response status code.