Network.AWS.LexModels.GetBotVersions

Description

Gets information about all of the versions of a bot.

The GetBotVersions operation returns a BotMetadata object for each version of a bot. For example, if a bot has three numbered versions, the GetBotVersions operation returns four BotMetadata objects in the response, one for each numbered version and one for the > LATEST version.

The GetBotVersions operation always returns at least one version, the > LATEST version.

This operation requires permissions for the lex:GetBotVersions action.

Creating a Request

getBotVersions

Arguments

:: Text
gbvName

-> GetBotVersions

Creates a value of GetBotVersions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbvNextToken - A pagination token for fetching the next page of bot versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

gbvMaxResults - The maximum number of bot versions to return in the response. The default is 10.

gbvName - The name of the bot for which versions should be returned.

data GetBotVersions

See: getBotVersions smart constructor.

Request Lenses

gbvNextToken :: Lens' GetBotVersions (Maybe Text)

A pagination token for fetching the next page of bot versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

gbvMaxResults :: Lens' GetBotVersions (Maybe Natural)

The maximum number of bot versions to return in the response. The default is 10.

gbvName :: Lens' GetBotVersions Text

The name of the bot for which versions should be returned.

Destructuring the Response

getBotVersionsResponse

Arguments

:: Int
gbvrsResponseStatus

-> GetBotVersionsResponse

Creates a value of GetBotVersionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbvrsBots - An array of BotMetadata objects, one for each numbered version of the bot plus one for the > LATEST version.

gbvrsNextToken - A pagination token for fetching the next page of bot versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

gbvrsResponseStatus - -- | The response status code.

data GetBotVersionsResponse

See: getBotVersionsResponse smart constructor.

Response Lenses

gbvrsBots :: Lens' GetBotVersionsResponse [BotMetadata]

An array of BotMetadata objects, one for each numbered version of the bot plus one for the > LATEST version.

gbvrsNextToken :: Lens' GetBotVersionsResponse (Maybe Text)

A pagination token for fetching the next page of bot versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

gbvrsResponseStatus :: Lens' GetBotVersionsResponse Int

- | The response status code.