Network.AWS.LexModels.GetBots

Description

Returns bot information as follows:

If you provide the nameContains field, the response includes information for the > LATEST version of all bots whose name contains the specified string.

If you don't specify the nameContains field, the operation returns information about the > LATEST version of all of your bots.

This operation requires permission for the lex:GetBots action.

Creating a Request

getBots :: GetBots

Creates a value of GetBots with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbNameContains - Substring to match in bot names. A bot will be returned if any part of its name matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz."

gbNextToken - A pagination token that fetches the next page of bots. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of bots, specify the pagination token in the next request.

gbMaxResults - The maximum number of bots to return in the response that the request will return. The default is 10.

data GetBots

See: getBots smart constructor.

Request Lenses

gbNameContains :: Lens' GetBots (Maybe Text)

Substring to match in bot names. A bot will be returned if any part of its name matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz."

gbNextToken :: Lens' GetBots (Maybe Text)

A pagination token that fetches the next page of bots. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of bots, specify the pagination token in the next request.

gbMaxResults :: Lens' GetBots (Maybe Natural)

The maximum number of bots to return in the response that the request will return. The default is 10.

Destructuring the Response

getBotsResponse

Arguments

:: Int
gbsrsResponseStatus

-> GetBotsResponse

Creates a value of GetBotsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbsrsBots - An array of botMetadata objects, with one entry for each bot.

gbsrsNextToken - If the response is truncated, it includes a pagination token that you can specify in your next request to fetch the next page of bots.

gbsrsResponseStatus - -- | The response status code.

data GetBotsResponse

See: getBotsResponse smart constructor.

Response Lenses

gbsrsBots :: Lens' GetBotsResponse [BotMetadata]

An array of botMetadata objects, with one entry for each bot.

gbsrsNextToken :: Lens' GetBotsResponse (Maybe Text)

If the response is truncated, it includes a pagination token that you can specify in your next request to fetch the next page of bots.

gbsrsResponseStatus :: Lens' GetBotsResponse Int

- | The response status code.