Network.AWS.LexModels.GetBuiltinIntent

Description

Returns information about a built-in intent.

This operation requires permission for the lex:GetBuiltinIntent action.

Creating a Request

getBuiltinIntent

Arguments

:: Text
gbiSignature

-> GetBuiltinIntent

Creates a value of GetBuiltinIntent with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbiSignature - The unique identifier for a built-in intent. To find the signature for an intent, see Standard Built-in Intents in the Alexa Skills Kit .

data GetBuiltinIntent

See: getBuiltinIntent smart constructor.

Request Lenses

gbiSignature :: Lens' GetBuiltinIntent Text

The unique identifier for a built-in intent. To find the signature for an intent, see Standard Built-in Intents in the Alexa Skills Kit .

Destructuring the Response

getBuiltinIntentResponse

Arguments

:: Int
gbirsResponseStatus

-> GetBuiltinIntentResponse

Creates a value of GetBuiltinIntentResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbirsSignature - The unique identifier for a built-in intent.

gbirsSlots - An array of BuiltinIntentSlot objects, one entry for each slot type in the intent.

gbirsSupportedLocales - A list of locales that the intent supports.

gbirsResponseStatus - -- | The response status code.

data GetBuiltinIntentResponse

See: getBuiltinIntentResponse smart constructor.

Response Lenses

gbirsSignature :: Lens' GetBuiltinIntentResponse (Maybe Text)

The unique identifier for a built-in intent.

gbirsSlots :: Lens' GetBuiltinIntentResponse [BuiltinIntentSlot]

An array of BuiltinIntentSlot objects, one entry for each slot type in the intent.

gbirsSupportedLocales :: Lens' GetBuiltinIntentResponse [Locale]

A list of locales that the intent supports.

gbirsResponseStatus :: Lens' GetBuiltinIntentResponse Int

- | The response status code.