Network.AWS.LexModels.GetBuiltinSlotTypes

Description

Gets a list of built-in slot types that meet the specified criteria.

For a list of built-in slot types, see Slot Type Reference in the Alexa Skills Kit .

This operation requires permission for the lex:GetBuiltInSlotTypes action.

Creating a Request

getBuiltinSlotTypes :: GetBuiltinSlotTypes

Creates a value of GetBuiltinSlotTypes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbstLocale - A list of locales that the slot type supports.

gbstNextToken - A pagination token that fetches the next page of slot types. If the response to this API call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of slot types, specify the pagination token in the next request.

gbstSignatureContains - Substring to match in built-in slot type signatures. A slot type will be returned if any part of its signature matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz."

gbstMaxResults - The maximum number of slot types to return in the response. The default is 10.

data GetBuiltinSlotTypes

See: getBuiltinSlotTypes smart constructor.

Request Lenses

gbstLocale :: Lens' GetBuiltinSlotTypes (Maybe Locale)

A list of locales that the slot type supports.

gbstNextToken :: Lens' GetBuiltinSlotTypes (Maybe Text)

A pagination token that fetches the next page of slot types. If the response to this API call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of slot types, specify the pagination token in the next request.

gbstSignatureContains :: Lens' GetBuiltinSlotTypes (Maybe Text)

Substring to match in built-in slot type signatures. A slot type will be returned if any part of its signature matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz."

gbstMaxResults :: Lens' GetBuiltinSlotTypes (Maybe Natural)

The maximum number of slot types to return in the response. The default is 10.

Destructuring the Response

getBuiltinSlotTypesResponse

Arguments

:: Int
gbstrsResponseStatus

-> GetBuiltinSlotTypesResponse

Creates a value of GetBuiltinSlotTypesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbstrsNextToken - If the response is truncated, the response includes a pagination token that you can use in your next request to fetch the next page of slot types.

gbstrsSlotTypes - An array of BuiltInSlotTypeMetadata objects, one entry for each slot type returned.

gbstrsResponseStatus - -- | The response status code.

data GetBuiltinSlotTypesResponse

See: getBuiltinSlotTypesResponse smart constructor.

Response Lenses

gbstrsNextToken :: Lens' GetBuiltinSlotTypesResponse (Maybe Text)

If the response is truncated, the response includes a pagination token that you can use in your next request to fetch the next page of slot types.

gbstrsSlotTypes :: Lens' GetBuiltinSlotTypesResponse [BuiltinSlotTypeMetadata]

An array of BuiltInSlotTypeMetadata objects, one entry for each slot type returned.

gbstrsResponseStatus :: Lens' GetBuiltinSlotTypesResponse Int

- | The response status code.