Network.AWS.LexModels.GetIntent

Description

Returns information about an intent. In addition to the intent name, you must specify the intent version.

This operation requires permissions to perform the lex:GetIntent action.

Creating a Request

getIntent

Arguments

:: Text
giName

-> Text
giVersion

-> GetIntent

Creates a value of GetIntent with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

giName - The name of the intent. The name is case sensitive.

giVersion - The version of the intent.

data GetIntent

See: getIntent smart constructor.

Request Lenses

giName :: Lens' GetIntent Text

The name of the intent. The name is case sensitive.

giVersion :: Lens' GetIntent Text

The version of the intent.

Destructuring the Response

getIntentResponse

Arguments

:: Int
girsResponseStatus

-> GetIntentResponse

Creates a value of GetIntentResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

girsFulfillmentActivity - Describes how the intent is fulfilled. For more information, see PutIntent .

girsSlots - An array of intent slots configured for the intent.

girsRejectionStatement - If the user answers "no" to the question defined in confirmationPrompt , Amazon Lex responds with this statement to acknowledge that the intent was canceled.

girsChecksum - Checksum of the intent.

girsConclusionStatement - After the Lambda function specified in the fulfillmentActivity element fulfills the intent, Amazon Lex conveys this statement to the user.

girsSampleUtterances - An array of sample utterances configured for the intent.

girsParentIntentSignature - A unique identifier for a built-in intent.

girsCreatedDate - The date that the intent was created.

girsName - The name of the intent.

girsVersion - The version of the intent.

girsFollowUpPrompt - If defined in the bot, Amazon Lex uses this prompt to solicit additional user activity after the intent is fulfilled. For more information, see PutIntent .

girsLastUpdatedDate - The date that the intent was updated. When you create a resource, the creation date and the last updated date are the same.

girsConfirmationPrompt - If defined in the bot, Amazon Lex uses prompt to confirm the intent before fulfilling the user's request. For more information, see PutIntent .

girsDialogCodeHook - If defined in the bot, Amazon Amazon Lex invokes this Lambda function for each user input. For more information, see PutIntent .

girsDescription - A description of the intent.

girsResponseStatus - -- | The response status code.

data GetIntentResponse

See: getIntentResponse smart constructor.

Response Lenses

girsFulfillmentActivity :: Lens' GetIntentResponse (Maybe FulfillmentActivity)

Describes how the intent is fulfilled. For more information, see PutIntent .

girsSlots :: Lens' GetIntentResponse [Slot]

An array of intent slots configured for the intent.

girsRejectionStatement :: Lens' GetIntentResponse (Maybe Statement)

If the user answers "no" to the question defined in confirmationPrompt , Amazon Lex responds with this statement to acknowledge that the intent was canceled.

girsChecksum :: Lens' GetIntentResponse (Maybe Text)

Checksum of the intent.

girsConclusionStatement :: Lens' GetIntentResponse (Maybe Statement)

After the Lambda function specified in the fulfillmentActivity element fulfills the intent, Amazon Lex conveys this statement to the user.

girsSampleUtterances :: Lens' GetIntentResponse [Text]

An array of sample utterances configured for the intent.

girsParentIntentSignature :: Lens' GetIntentResponse (Maybe Text)

A unique identifier for a built-in intent.

girsCreatedDate :: Lens' GetIntentResponse (Maybe UTCTime)

The date that the intent was created.

girsName :: Lens' GetIntentResponse (Maybe Text)

The name of the intent.

girsVersion :: Lens' GetIntentResponse (Maybe Text)

The version of the intent.

girsFollowUpPrompt :: Lens' GetIntentResponse (Maybe FollowUpPrompt)

If defined in the bot, Amazon Lex uses this prompt to solicit additional user activity after the intent is fulfilled. For more information, see PutIntent .

girsLastUpdatedDate :: Lens' GetIntentResponse (Maybe UTCTime)

The date that the intent was updated. When you create a resource, the creation date and the last updated date are the same.

girsConfirmationPrompt :: Lens' GetIntentResponse (Maybe Prompt)

If defined in the bot, Amazon Lex uses prompt to confirm the intent before fulfilling the user's request. For more information, see PutIntent .

girsDialogCodeHook :: Lens' GetIntentResponse (Maybe CodeHook)

If defined in the bot, Amazon Amazon Lex invokes this Lambda function for each user input. For more information, see PutIntent .

girsDescription :: Lens' GetIntentResponse (Maybe Text)

A description of the intent.

girsResponseStatus :: Lens' GetIntentResponse Int

- | The response status code.