Network.AWS.LexModels.GetIntents

Description

Returns intent information as follows:

If you specify the nameContains field, returns the > LATEST version of all intents that contain the specified string.

If you don't specify the nameContains field, returns information about the > LATEST version of all intents.

The operation requires permission for the lex:GetIntents action.

Creating a Request

getIntents :: GetIntents

Creates a value of GetIntents with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

giNameContains - Substring to match in intent names. An intent will be returned if any part of its name matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz."

giNextToken - A pagination token that fetches the next page of intents. If the response to this API call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of intents, specify the pagination token in the next request.

giMaxResults - The maximum number of intents to return in the response. The default is 10.

data GetIntents

See: getIntents smart constructor.

Request Lenses

giNameContains :: Lens' GetIntents (Maybe Text)

Substring to match in intent names. An intent will be returned if any part of its name matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz."

giNextToken :: Lens' GetIntents (Maybe Text)

A pagination token that fetches the next page of intents. If the response to this API call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of intents, specify the pagination token in the next request.

giMaxResults :: Lens' GetIntents (Maybe Natural)

The maximum number of intents to return in the response. The default is 10.

Destructuring the Response

getIntentsResponse

Arguments

:: Int
gisrsResponseStatus

-> GetIntentsResponse

Creates a value of GetIntentsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gisrsIntents - An array of Intent objects. For more information, see PutBot .

gisrsNextToken - If the response is truncated, the response includes a pagination token that you can specify in your next request to fetch the next page of intents.

gisrsResponseStatus - -- | The response status code.

data GetIntentsResponse

See: getIntentsResponse smart constructor.

Response Lenses

gisrsIntents :: Lens' GetIntentsResponse [IntentMetadata]

An array of Intent objects. For more information, see PutBot .

gisrsNextToken :: Lens' GetIntentsResponse (Maybe Text)

If the response is truncated, the response includes a pagination token that you can specify in your next request to fetch the next page of intents.

gisrsResponseStatus :: Lens' GetIntentsResponse Int

- | The response status code.