Network.AWS.LexModels.GetSlotType

Description

Returns information about a specific version of a slot type. In addition to specifying the slot type name, you must specify the slot type version.

This operation requires permissions for the lex:GetSlotType action.

Creating a Request

getSlotType

Arguments

:: Text
gstName

-> Text
gstVersion

-> GetSlotType

Creates a value of GetSlotType with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gstName - The name of the slot type. The name is case sensitive.

gstVersion - The version of the slot type.

data GetSlotType

See: getSlotType smart constructor.

Request Lenses

gstName :: Lens' GetSlotType Text

The name of the slot type. The name is case sensitive.

gstVersion :: Lens' GetSlotType Text

The version of the slot type.

Destructuring the Response

getSlotTypeResponse

Arguments

:: Int
getrsResponseStatus

-> GetSlotTypeResponse

Creates a value of GetSlotTypeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

getrsChecksum - Checksum of the > LATEST version of the slot type.

getrsCreatedDate - The date that the slot type was created.

getrsName - The name of the slot type.

getrsVersion - The version of the slot type.

getrsLastUpdatedDate - The date that the slot type was updated. When you create a resource, the creation date and last update date are the same.

getrsDescription - A description of the slot type.

getrsEnumerationValues - A list of EnumerationValue objects that defines the values that the slot type can take.

getrsResponseStatus - -- | The response status code.

data GetSlotTypeResponse

See: getSlotTypeResponse smart constructor.

Response Lenses

getrsChecksum :: Lens' GetSlotTypeResponse (Maybe Text)

Checksum of the > LATEST version of the slot type.

getrsCreatedDate :: Lens' GetSlotTypeResponse (Maybe UTCTime)

The date that the slot type was created.

getrsName :: Lens' GetSlotTypeResponse (Maybe Text)

The name of the slot type.

getrsVersion :: Lens' GetSlotTypeResponse (Maybe Text)

The version of the slot type.

getrsLastUpdatedDate :: Lens' GetSlotTypeResponse (Maybe UTCTime)

The date that the slot type was updated. When you create a resource, the creation date and last update date are the same.

getrsDescription :: Lens' GetSlotTypeResponse (Maybe Text)

A description of the slot type.

getrsEnumerationValues :: Lens' GetSlotTypeResponse (Maybe (NonEmpty EnumerationValue))

A list of EnumerationValue objects that defines the values that the slot type can take.

getrsResponseStatus :: Lens' GetSlotTypeResponse Int

- | The response status code.