Network.AWS.LexModels.GetSlotTypeVersions

Description

Gets information about all versions of a slot type.

The GetSlotTypeVersions operation returns a SlotTypeMetadata object for each version of a slot type. For example, if a slot type has three numbered versions, the GetSlotTypeVersions operation returns four SlotTypeMetadata objects in the response, one for each numbered version and one for the > LATEST version.

The GetSlotTypeVersions operation always returns at least one version, the > LATEST version.

This operation requires permissions for the lex:GetSlotTypeVersions action.

Creating a Request

getSlotTypeVersions

Arguments

:: Text
gstvName

-> GetSlotTypeVersions

Creates a value of GetSlotTypeVersions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gstvNextToken - A pagination token for fetching the next page of slot type versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

gstvMaxResults - The maximum number of slot type versions to return in the response. The default is 10.

gstvName - The name of the slot type for which versions should be returned.

data GetSlotTypeVersions

See: getSlotTypeVersions smart constructor.

Request Lenses

gstvNextToken :: Lens' GetSlotTypeVersions (Maybe Text)

A pagination token for fetching the next page of slot type versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

gstvMaxResults :: Lens' GetSlotTypeVersions (Maybe Natural)

The maximum number of slot type versions to return in the response. The default is 10.

gstvName :: Lens' GetSlotTypeVersions Text

The name of the slot type for which versions should be returned.

Destructuring the Response

getSlotTypeVersionsResponse

Arguments

:: Int
gstvrsResponseStatus

-> GetSlotTypeVersionsResponse

Creates a value of GetSlotTypeVersionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gstvrsNextToken - A pagination token for fetching the next page of slot type versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

gstvrsSlotTypes - An array of SlotTypeMetadata objects, one for each numbered version of the slot type plus one for the > LATEST version.

gstvrsResponseStatus - -- | The response status code.

data GetSlotTypeVersionsResponse

See: getSlotTypeVersionsResponse smart constructor.

Response Lenses

gstvrsNextToken :: Lens' GetSlotTypeVersionsResponse (Maybe Text)

A pagination token for fetching the next page of slot type versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

gstvrsSlotTypes :: Lens' GetSlotTypeVersionsResponse [SlotTypeMetadata]

An array of SlotTypeMetadata objects, one for each numbered version of the slot type plus one for the > LATEST version.

gstvrsResponseStatus :: Lens' GetSlotTypeVersionsResponse Int

- | The response status code.