Network.AWS.LexModels.GetUtterancesView

Description

Use the GetUtterancesView operation to get information about the utterances that your users have made to your bot. You can use this list to tune the utterances that your bot responds to.

For example, say that you have created a bot to order flowers. After your users have used your bot for a while, use the GetUtterancesView operation to see the requests that they have made and whether they have been successful. You might find that the utterance "I want flowers" is not being recognized. You could add this utterance to the OrderFlowers intent so that your bot recognizes that utterance.

After you publish a new version of a bot, you can get information about the old version and the new so that you can compare the performance across the two versions.

Data is available for the last 15 days. You can request information for up to 5 versions in each request. The response contains information about a maximum of 100 utterances for each version.

If the bot's childDirected field is set to true , utterances for the bot are not stored and cannot be retrieved with the GetUtterancesView operation. For more information, see PutBot .

This operation requires permissions for the lex:GetUtterancesView action.

Creating a Request

getUtterancesView

Arguments

:: Text
guvBotName

-> NonEmpty Text
guvBotVersions

-> StatusType
guvStatusType

-> GetUtterancesView

Creates a value of GetUtterancesView with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

guvBotName - The name of the bot for which utterance information should be returned.

guvBotVersions - An array of bot versions for which utterance information should be returned. The limit is 5 versions per request.

guvStatusType - To return utterances that were recognized and handled, useDetected . To return utterances that were not recognized, use Missed .

data GetUtterancesView

See: getUtterancesView smart constructor.

Request Lenses

guvBotName :: Lens' GetUtterancesView Text

The name of the bot for which utterance information should be returned.

guvBotVersions :: Lens' GetUtterancesView (NonEmpty Text)

An array of bot versions for which utterance information should be returned. The limit is 5 versions per request.

guvStatusType :: Lens' GetUtterancesView StatusType

To return utterances that were recognized and handled, useDetected . To return utterances that were not recognized, use Missed .

Destructuring the Response

getUtterancesViewResponse

Arguments

:: Int
guvrsResponseStatus

-> GetUtterancesViewResponse

Creates a value of GetUtterancesViewResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

guvrsBotName - The name of the bot for which utterance information was returned.

guvrsUtterances - An array of UtteranceList objects, each containing a list of UtteranceData objects describing the utterances that were processed by your bot. The response contains a maximum of 100 UtteranceData objects for each version.

guvrsResponseStatus - -- | The response status code.

data GetUtterancesViewResponse

See: getUtterancesViewResponse smart constructor.

Response Lenses

guvrsBotName :: Lens' GetUtterancesViewResponse (Maybe Text)

The name of the bot for which utterance information was returned.

guvrsUtterances :: Lens' GetUtterancesViewResponse [UtteranceList]

An array of UtteranceList objects, each containing a list of UtteranceData objects describing the utterances that were processed by your bot. The response contains a maximum of 100 UtteranceData objects for each version.

guvrsResponseStatus :: Lens' GetUtterancesViewResponse Int

- | The response status code.