Network.AWS.LexModels.PutBot

Description

Creates an Amazon Lex conversational bot or replaces an existing bot. When you create or update a bot you are only required to specify a name. You can use this to add intents later, or to remove intents from an existing bot. When you create a bot with a name only, the bot is created or updated but Amazon Lex returns the response FAILED . You can build the bot after you add one or more intents. For more information about Amazon Lex bots, see 'how-it-works' .

If you specify the name of an existing bot, the fields in the request replace the existing values in the > LATEST version of the bot. Amazon Lex removes any fields that you don't provide values for in the request, except for the idleTTLInSeconds and privacySettings fields, which are set to their default values. If you don't specify values for required fields, Amazon Lex throws an exception.

This operation requires permissions for the lex:PutBot action. For more information, see 'auth-and-access-control' .

Creating a Request

putBot

Arguments

:: Text
pbName

-> Locale
pbLocale

-> Bool
pbChildDirected

-> PutBot

Creates a value of PutBot with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pbAbortStatement - When Amazon Lex can't understand the user's input in context, it tries to elicit the information a few times. After that, Amazon Lex sends the message defined in abortStatement to the user, and then aborts the conversation. To set the number of retries, use the valueElicitationPrompt field for the slot type. For example, in a pizza ordering bot, Amazon Lex might ask a user "What type of crust would you like?" If the user's response is not one of the expected responses (for example, "thin crust, "deep dish," etc.), Amazon Lex tries to elicit a correct response a few more times. For example, in a pizza ordering application, OrderPizza might be one of the intents. This intent might require the CrustType slot. You specify the valueElicitationPrompt field when you create the CrustType slot.

pbIntents - An array of Intent objects. Each intent represents a command that a user can express. For example, a pizza ordering bot might support an OrderPizza intent. For more information, see 'how-it-works' .

pbChecksum - Identifies a specific revision of the > LATEST version. When you create a new bot, leave the checksum field blank. If you specify a checksum you get a BadRequestException exception. When you want to update a bot, set the checksum field to the checksum of the most recent revision of the > LATEST version. If you don't specify the checksum field, or if the checksum does not match the > LATEST version, you get a PreconditionFailedException exception.

pbProcessBehavior - If you set the processBehavior element to Build , Amazon Lex builds the bot so that it can be run. If you set the element to Save Amazon Lex saves the bot, but doesn't build it. If you don't specify this value, the default value is Save .

pbIdleSessionTTLInSeconds - The maximum time in seconds that Amazon Lex retains the data gathered in a conversation. A user interaction session remains active for the amount of time specified. If no conversation occurs during this time, the session expires and Amazon Lex deletes any data provided before the timeout. For example, suppose that a user chooses the OrderPizza intent, but gets sidetracked halfway through placing an order. If the user doesn't complete the order within the specified time, Amazon Lex discards the slot information that it gathered, and the user must start over. If you don't include the idleSessionTTLInSeconds element in a PutBot operation request, Amazon Lex uses the default value. This is also true if the request replaces an existing bot. The default is 300 seconds (5 minutes).

pbClarificationPrompt - When Amazon Lex doesn't understand the user's intent, it uses this message to get clarification. To specify how many times Amazon Lex should repeate the clarification prompt, use the maxAttempts field. If Amazon Lex still doesn't understand, it sends the message in the abortStatement field. When you create a clarification prompt, make sure that it suggests the correct response from the user. for example, for a bot that orders pizza and drinks, you might create this clarification prompt: "What would you like to do? You can say 'Order a pizza' or 'Order a drink.'"

pbVoiceId - The Amazon Polly voice ID that you want Amazon Lex to use for voice interactions with the user. The locale configured for the voice must match the locale of the bot. For more information, see Available Voices in the Amazon Polly Developer Guide .

pbDescription - A description of the bot.

pbName - The name of the bot. The name is not case sensitive.

pbLocale - Specifies the target locale for the bot. Any intent used in the bot must be compatible with the locale of the bot. The default is en-US .

pbChildDirected - For each Amazon Lex bot created with the Amazon Lex Model Building Service, you must specify whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to the Children's Online Privacy Protection Act (COPPA) by specifying true or false in the childDirected field. By specifying true in the childDirected field, you confirm that your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. By specifying false in the childDirected field, you confirm that your use of Amazon Lex is not related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. You may not specify a default value for the childDirected field that does not accurately reflect whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. If your use of Amazon Lex relates to a website, program, or other application that is directed in whole or in part, to children under age 13, you must obtain any required verifiable parental consent under COPPA. For information regarding the use of Amazon Lex in connection with websites, programs, or other applications that are directed or targeted, in whole or in part, to children under age 13, see the Amazon Lex FAQ.

data PutBot

See: putBot smart constructor.

Request Lenses

pbAbortStatement :: Lens' PutBot (Maybe Statement)

When Amazon Lex can't understand the user's input in context, it tries to elicit the information a few times. After that, Amazon Lex sends the message defined in abortStatement to the user, and then aborts the conversation. To set the number of retries, use the valueElicitationPrompt field for the slot type. For example, in a pizza ordering bot, Amazon Lex might ask a user "What type of crust would you like?" If the user's response is not one of the expected responses (for example, "thin crust, "deep dish," etc.), Amazon Lex tries to elicit a correct response a few more times. For example, in a pizza ordering application, OrderPizza might be one of the intents. This intent might require the CrustType slot. You specify the valueElicitationPrompt field when you create the CrustType slot.

pbIntents :: Lens' PutBot (Maybe (NonEmpty Intent))

An array of Intent objects. Each intent represents a command that a user can express. For example, a pizza ordering bot might support an OrderPizza intent. For more information, see 'how-it-works' .

pbChecksum :: Lens' PutBot (Maybe Text)

Identifies a specific revision of the > LATEST version. When you create a new bot, leave the checksum field blank. If you specify a checksum you get a BadRequestException exception. When you want to update a bot, set the checksum field to the checksum of the most recent revision of the > LATEST version. If you don't specify the checksum field, or if the checksum does not match the > LATEST version, you get a PreconditionFailedException exception.

pbProcessBehavior :: Lens' PutBot (Maybe ProcessBehavior)

If you set the processBehavior element to Build , Amazon Lex builds the bot so that it can be run. If you set the element to Save Amazon Lex saves the bot, but doesn't build it. If you don't specify this value, the default value is Save .

pbIdleSessionTTLInSeconds :: Lens' PutBot (Maybe Natural)

The maximum time in seconds that Amazon Lex retains the data gathered in a conversation. A user interaction session remains active for the amount of time specified. If no conversation occurs during this time, the session expires and Amazon Lex deletes any data provided before the timeout. For example, suppose that a user chooses the OrderPizza intent, but gets sidetracked halfway through placing an order. If the user doesn't complete the order within the specified time, Amazon Lex discards the slot information that it gathered, and the user must start over. If you don't include the idleSessionTTLInSeconds element in a PutBot operation request, Amazon Lex uses the default value. This is also true if the request replaces an existing bot. The default is 300 seconds (5 minutes).

pbClarificationPrompt :: Lens' PutBot (Maybe Prompt)

When Amazon Lex doesn't understand the user's intent, it uses this message to get clarification. To specify how many times Amazon Lex should repeate the clarification prompt, use the maxAttempts field. If Amazon Lex still doesn't understand, it sends the message in the abortStatement field. When you create a clarification prompt, make sure that it suggests the correct response from the user. for example, for a bot that orders pizza and drinks, you might create this clarification prompt: "What would you like to do? You can say 'Order a pizza' or 'Order a drink.'"

pbVoiceId :: Lens' PutBot (Maybe Text)

The Amazon Polly voice ID that you want Amazon Lex to use for voice interactions with the user. The locale configured for the voice must match the locale of the bot. For more information, see Available Voices in the Amazon Polly Developer Guide .

pbDescription :: Lens' PutBot (Maybe Text)

A description of the bot.

pbName :: Lens' PutBot Text

The name of the bot. The name is not case sensitive.

pbLocale :: Lens' PutBot Locale

Specifies the target locale for the bot. Any intent used in the bot must be compatible with the locale of the bot. The default is en-US .

pbChildDirected :: Lens' PutBot Bool

For each Amazon Lex bot created with the Amazon Lex Model Building Service, you must specify whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to the Children's Online Privacy Protection Act (COPPA) by specifying true or false in the childDirected field. By specifying true in the childDirected field, you confirm that your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. By specifying false in the childDirected field, you confirm that your use of Amazon Lex is not related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. You may not specify a default value for the childDirected field that does not accurately reflect whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. If your use of Amazon Lex relates to a website, program, or other application that is directed in whole or in part, to children under age 13, you must obtain any required verifiable parental consent under COPPA. For information regarding the use of Amazon Lex in connection with websites, programs, or other applications that are directed or targeted, in whole or in part, to children under age 13, see the Amazon Lex FAQ.

Destructuring the Response

putBotResponse

Arguments

:: Int
pbrsResponseStatus

-> PutBotResponse

Creates a value of PutBotResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pbrsFailureReason - If status is FAILED , Amazon Lex provides the reason that it failed to build the bot.

pbrsStatus - When you send a request to create a bot with processBehavior set to BUILD , Amazon Lex sets the status response element to BUILDING . After Amazon Lex builds the bot, it sets status to READY . If Amazon Lex can't build the bot, Amazon Lex sets status to FAILED . Amazon Lex returns the reason for the failure in the failureReason response element. When you set processBehavior to SAVE , Amazon Lex sets the status code to NOT BUILT .

pbrsAbortStatement - The message that Amazon Lex uses to abort a conversation. For more information, see PutBot .

pbrsIntents - An array of Intent objects. For more information, see PutBot .

pbrsChecksum - Checksum of the bot that you created.

pbrsLocale - The target locale for the bot.

pbrsCreatedDate - The date that the bot was created.

pbrsName - The name of the bot.

pbrsVersion - The version of the bot. For a new bot, the version is always > LATEST .

pbrsIdleSessionTTLInSeconds - The maximum length of time that Amazon Lex retains the data gathered in a conversation. For more information, see PutBot .

pbrsClarificationPrompt - The prompts that Amazon Lex uses when it doesn't understand the user's intent. For more information, see PutBot .

pbrsVoiceId - The Amazon Polly voice ID that Amazon Lex uses for voice interaction with the user. For more information, see PutBot .

pbrsLastUpdatedDate - The date that the bot was updated. When you create a resource, the creation date and last updated date are the same.

pbrsChildDirected - For each Amazon Lex bot created with the Amazon Lex Model Building Service, you must specify whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to the Children's Online Privacy Protection Act (COPPA) by specifying true or false in the childDirected field. By specifying true in the childDirected field, you confirm that your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. By specifying false in the childDirected field, you confirm that your use of Amazon Lex is not related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. You may not specify a default value for the childDirected field that does not accurately reflect whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. If your use of Amazon Lex relates to a website, program, or other application that is directed in whole or in part, to children under age 13, you must obtain any required verifiable parental consent under COPPA. For information regarding the use of Amazon Lex in connection with websites, programs, or other applications that are directed or targeted, in whole or in part, to children under age 13, see the Amazon Lex FAQ.

pbrsDescription - A description of the bot.

pbrsResponseStatus - -- | The response status code.

data PutBotResponse

See: putBotResponse smart constructor.

Response Lenses

pbrsFailureReason :: Lens' PutBotResponse (Maybe Text)

If status is FAILED , Amazon Lex provides the reason that it failed to build the bot.

pbrsStatus :: Lens' PutBotResponse (Maybe LexStatus)

When you send a request to create a bot with processBehavior set to BUILD , Amazon Lex sets the status response element to BUILDING . After Amazon Lex builds the bot, it sets status to READY . If Amazon Lex can't build the bot, Amazon Lex sets status to FAILED . Amazon Lex returns the reason for the failure in the failureReason response element. When you set processBehavior to SAVE , Amazon Lex sets the status code to NOT BUILT .

pbrsAbortStatement :: Lens' PutBotResponse (Maybe Statement)

The message that Amazon Lex uses to abort a conversation. For more information, see PutBot .

pbrsIntents :: Lens' PutBotResponse (Maybe (NonEmpty Intent))

An array of Intent objects. For more information, see PutBot .

pbrsChecksum :: Lens' PutBotResponse (Maybe Text)

Checksum of the bot that you created.

pbrsLocale :: Lens' PutBotResponse (Maybe Locale)

The target locale for the bot.

pbrsCreatedDate :: Lens' PutBotResponse (Maybe UTCTime)

The date that the bot was created.

pbrsName :: Lens' PutBotResponse (Maybe Text)

The name of the bot.

pbrsVersion :: Lens' PutBotResponse (Maybe Text)

The version of the bot. For a new bot, the version is always > LATEST .

pbrsIdleSessionTTLInSeconds :: Lens' PutBotResponse (Maybe Natural)

The maximum length of time that Amazon Lex retains the data gathered in a conversation. For more information, see PutBot .

pbrsClarificationPrompt :: Lens' PutBotResponse (Maybe Prompt)

The prompts that Amazon Lex uses when it doesn't understand the user's intent. For more information, see PutBot .

pbrsVoiceId :: Lens' PutBotResponse (Maybe Text)

The Amazon Polly voice ID that Amazon Lex uses for voice interaction with the user. For more information, see PutBot .

pbrsLastUpdatedDate :: Lens' PutBotResponse (Maybe UTCTime)

The date that the bot was updated. When you create a resource, the creation date and last updated date are the same.

pbrsChildDirected :: Lens' PutBotResponse (Maybe Bool)

For each Amazon Lex bot created with the Amazon Lex Model Building Service, you must specify whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to the Children's Online Privacy Protection Act (COPPA) by specifying true or false in the childDirected field. By specifying true in the childDirected field, you confirm that your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. By specifying false in the childDirected field, you confirm that your use of Amazon Lex is not related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. You may not specify a default value for the childDirected field that does not accurately reflect whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. If your use of Amazon Lex relates to a website, program, or other application that is directed in whole or in part, to children under age 13, you must obtain any required verifiable parental consent under COPPA. For information regarding the use of Amazon Lex in connection with websites, programs, or other applications that are directed or targeted, in whole or in part, to children under age 13, see the Amazon Lex FAQ.

pbrsDescription :: Lens' PutBotResponse (Maybe Text)

A description of the bot.

pbrsResponseStatus :: Lens' PutBotResponse Int

- | The response status code.