Network.AWS.LexModels.PutBotAlias

Description

Creates an alias for the specified version of the bot or replaces an alias for the specified bot. To change the version of the bot that the alias points to, replace the alias. For more information about aliases, see 'versioning-aliases' .

This operation requires permissions for the lex:PutBotAlias action.

Creating a Request

putBotAlias

Arguments

:: Text
pbaName

-> Text
pbaBotVersion

-> Text
pbaBotName

-> PutBotAlias

Creates a value of PutBotAlias with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pbaChecksum - Identifies a specific revision of the > LATEST version. When you create a new bot alias, leave the checksum field blank. If you specify a checksum you get a BadRequestException exception. When you want to update a bot alias, set the checksum field to the checksum of the most recent revision of the > LATEST version. If you don't specify the checksum field, or if the checksum does not match the > LATEST version, you get a PreconditionFailedException exception.

pbaDescription - A description of the alias.

pbaName - The name of the alias. The name is not case sensitive.

pbaBotVersion - The version of the bot.

pbaBotName - The name of the bot.

data PutBotAlias

See: putBotAlias smart constructor.

Request Lenses

pbaChecksum :: Lens' PutBotAlias (Maybe Text)

Identifies a specific revision of the > LATEST version. When you create a new bot alias, leave the checksum field blank. If you specify a checksum you get a BadRequestException exception. When you want to update a bot alias, set the checksum field to the checksum of the most recent revision of the > LATEST version. If you don't specify the checksum field, or if the checksum does not match the > LATEST version, you get a PreconditionFailedException exception.

pbaDescription :: Lens' PutBotAlias (Maybe Text)

A description of the alias.

pbaName :: Lens' PutBotAlias Text

The name of the alias. The name is not case sensitive.

pbaBotVersion :: Lens' PutBotAlias Text

The version of the bot.

pbaBotName :: Lens' PutBotAlias Text

The name of the bot.

Destructuring the Response

putBotAliasResponse

Arguments

:: Int
pbarsResponseStatus

-> PutBotAliasResponse

Creates a value of PutBotAliasResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pbarsChecksum - The checksum for the current version of the alias.

pbarsBotVersion - The version of the bot that the alias points to.

pbarsBotName - The name of the bot that the alias points to.

pbarsCreatedDate - The date that the bot alias was created.

pbarsName - The name of the alias.

pbarsLastUpdatedDate - The date that the bot alias was updated. When you create a resource, the creation date and the last updated date are the same.

pbarsDescription - A description of the alias.

pbarsResponseStatus - -- | The response status code.

data PutBotAliasResponse

See: putBotAliasResponse smart constructor.

Response Lenses

pbarsChecksum :: Lens' PutBotAliasResponse (Maybe Text)

The checksum for the current version of the alias.

pbarsBotVersion :: Lens' PutBotAliasResponse (Maybe Text)

The version of the bot that the alias points to.

pbarsBotName :: Lens' PutBotAliasResponse (Maybe Text)

The name of the bot that the alias points to.

pbarsCreatedDate :: Lens' PutBotAliasResponse (Maybe UTCTime)

The date that the bot alias was created.

pbarsName :: Lens' PutBotAliasResponse (Maybe Text)

The name of the alias.

pbarsLastUpdatedDate :: Lens' PutBotAliasResponse (Maybe UTCTime)

The date that the bot alias was updated. When you create a resource, the creation date and the last updated date are the same.

pbarsDescription :: Lens' PutBotAliasResponse (Maybe Text)

A description of the alias.

pbarsResponseStatus :: Lens' PutBotAliasResponse Int

- | The response status code.