Network.AWS.LexModels.PutIntent

Description

Creates an intent or replaces an existing intent.

To define the interaction between the user and your bot, you use one or more intents. For a pizza ordering bot, for example, you would create an OrderPizza intent.

To create an intent or replace an existing intent, you must provide the following:

Intent name. For example, OrderPizza .

Sample utterances. For example, "Can I order a pizza, please." and "I want to order a pizza."

Information to be gathered. You specify slot types for the information that your bot will request from the user. You can specify standard slot types, such as a date or a time, or custom slot types such as the size and crust of a pizza.

How the intent will be fulfilled. You can provide a Lambda function or configure the intent to return the intent information to the client application. If you use a Lambda function, when all of the intent information is available, Amazon Lex invokes your Lambda function. If you configure your intent to return the intent information to the client application.

You can specify other optional information in the request, such as:

A confirmation prompt to ask the user to confirm an intent. For example, "Shall I order your pizza?"

A conclusion statement to send to the user after the intent has been fulfilled. For example, "I placed your pizza order."

A follow-up prompt that asks the user for additional activity. For example, asking "Do you want to order a drink with your pizza?"

If you specify an existing intent name to update the intent, Amazon Lex replaces the values in the > LATEST version of the slot type with the values in the request. Amazon Lex removes fields that you don't provide in the request. If you don't specify the required fields, Amazon Lex throws an exception.

For more information, see 'how-it-works' .

This operation requires permissions for the lex:PutIntent action.

Creating a Request

putIntent

Arguments

:: Text
piName

-> PutIntent

Creates a value of PutIntent with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

piFulfillmentActivity - Describes how the intent is fulfilled. For example, after a user provides all of the information for a pizza order, fulfillmentActivity defines how the bot places an order with a local pizza store. You might configure Amazon Lex to return all of the intent information to the client application, or direct it to invoke a Lambda function that can process the intent (for example, place an order with a pizzeria).

piSlots - An array of intent slots. At runtime, Amazon Lex elicits required slot values from the user using prompts defined in the slots. For more information, see linkend="how-it-works"/.

piRejectionStatement - When the user answers "no" to the question defined in confirmationPrompt , Amazon Lex responds with this statement to acknowledge that the intent was canceled.

piChecksum - Identifies a specific revision of the > LATEST version. When you create a new intent, leave the checksum field blank. If you specify a checksum you get a BadRequestException exception. When you want to update a intent, set the checksum field to the checksum of the most recent revision of the > LATEST version. If you don't specify the checksum field, or if the checksum does not match the > LATEST version, you get a PreconditionFailedException exception.

piConclusionStatement - The statement that you want Amazon Lex to convey to the user after the intent is successfully fulfilled by the Lambda function. This element is relevant only if you provide a Lambda function in the fulfillmentActivity . If you return the intent to the client application, you can't specify this element.

piSampleUtterances - An array of utterances (strings) that a user might say to signal the intent. For example, "I want {PizzaSize} pizza", "Order {Quantity} {PizzaSize} pizzas". In each utterance, a slot name is enclosed in curly braces.

piParentIntentSignature - A unique identifier for the built-in intent to base this intent on. To find the signature for an intent, see Standard Built-in Intents in the Alexa Skills Kit .

piFollowUpPrompt - Amazon Lex uses this prompt to solicit additional activity after fulfilling an intent. For example, after the OrderPizza intent is fulfilled, you might prompt the user to order a drink. The action that Amazon Lex takes depends on the user's response, as follows: * If the user says Yes it responds with the clarification prompt that is configured for the bot. * if the user says Yes and continues with an utterance that triggers an intent it starts a conversation for the intent. * If the user says No it responds with the rejection statement configured for the the follow-up prompt. * If it doesn't recognize the utterance it repeats the follow-up prompt again. The followUpPrompt field and the conclusionStatement field are mutually exclusive. You can specify only one.

piConfirmationPrompt - Prompts the user to confirm the intent. This question should have a yes or no answer. Amazon Lex uses this prompt to ensure that the user acknowledges that the intent is ready for fulfillment. For example, with the OrderPizza intent, you might want to confirm that the order is correct before placing it. For other intents, such as intents that simply respond to user questions, you might not need to ask the user for confirmation before providing the information.

piDialogCodeHook - Specifies a Lambda function to invoke for each user input. You can invoke this Lambda function to personalize user interaction. For example, suppose your bot determines that the user is John. Your Lambda function might retrieve John's information from a backend database and prepopulate some of the values. For example, if you find that John is gluten intolerant, you might set the corresponding intent slot, GlutenIntolerant , to true. You might find John's phone number and set the corresponding session attribute.

piDescription - A description of the intent.

piName - The name of the intent. The name is not case sensitive. The name can't match a built-in intent name, or a built-in intent name with "AMAZON." removed. For example, because there is a built-in intent called AMAZON.HelpIntent , you can't create a custom intent called HelpIntent . For a list of built-in intents, see Standard Built-in Intents in the Alexa Skills Kit .

data PutIntent

See: putIntent smart constructor.

Request Lenses

piFulfillmentActivity :: Lens' PutIntent (Maybe FulfillmentActivity)

Describes how the intent is fulfilled. For example, after a user provides all of the information for a pizza order, fulfillmentActivity defines how the bot places an order with a local pizza store. You might configure Amazon Lex to return all of the intent information to the client application, or direct it to invoke a Lambda function that can process the intent (for example, place an order with a pizzeria).

piSlots :: Lens' PutIntent [Slot]

An array of intent slots. At runtime, Amazon Lex elicits required slot values from the user using prompts defined in the slots. For more information, see linkend="how-it-works"/.

piRejectionStatement :: Lens' PutIntent (Maybe Statement)

When the user answers "no" to the question defined in confirmationPrompt , Amazon Lex responds with this statement to acknowledge that the intent was canceled.

piChecksum :: Lens' PutIntent (Maybe Text)

Identifies a specific revision of the > LATEST version. When you create a new intent, leave the checksum field blank. If you specify a checksum you get a BadRequestException exception. When you want to update a intent, set the checksum field to the checksum of the most recent revision of the > LATEST version. If you don't specify the checksum field, or if the checksum does not match the > LATEST version, you get a PreconditionFailedException exception.

piConclusionStatement :: Lens' PutIntent (Maybe Statement)

The statement that you want Amazon Lex to convey to the user after the intent is successfully fulfilled by the Lambda function. This element is relevant only if you provide a Lambda function in the fulfillmentActivity . If you return the intent to the client application, you can't specify this element.

piSampleUtterances :: Lens' PutIntent [Text]

An array of utterances (strings) that a user might say to signal the intent. For example, "I want {PizzaSize} pizza", "Order {Quantity} {PizzaSize} pizzas". In each utterance, a slot name is enclosed in curly braces.

piParentIntentSignature :: Lens' PutIntent (Maybe Text)

A unique identifier for the built-in intent to base this intent on. To find the signature for an intent, see Standard Built-in Intents in the Alexa Skills Kit .

piFollowUpPrompt :: Lens' PutIntent (Maybe FollowUpPrompt)

Amazon Lex uses this prompt to solicit additional activity after fulfilling an intent. For example, after the OrderPizza intent is fulfilled, you might prompt the user to order a drink. The action that Amazon Lex takes depends on the user's response, as follows: * If the user says Yes it responds with the clarification prompt that is configured for the bot. * if the user says Yes and continues with an utterance that triggers an intent it starts a conversation for the intent. * If the user says No it responds with the rejection statement configured for the the follow-up prompt. * If it doesn't recognize the utterance it repeats the follow-up prompt again. The followUpPrompt field and the conclusionStatement field are mutually exclusive. You can specify only one.

piConfirmationPrompt :: Lens' PutIntent (Maybe Prompt)

Prompts the user to confirm the intent. This question should have a yes or no answer. Amazon Lex uses this prompt to ensure that the user acknowledges that the intent is ready for fulfillment. For example, with the OrderPizza intent, you might want to confirm that the order is correct before placing it. For other intents, such as intents that simply respond to user questions, you might not need to ask the user for confirmation before providing the information.

piDialogCodeHook :: Lens' PutIntent (Maybe CodeHook)

Specifies a Lambda function to invoke for each user input. You can invoke this Lambda function to personalize user interaction. For example, suppose your bot determines that the user is John. Your Lambda function might retrieve John's information from a backend database and prepopulate some of the values. For example, if you find that John is gluten intolerant, you might set the corresponding intent slot, GlutenIntolerant , to true. You might find John's phone number and set the corresponding session attribute.

piDescription :: Lens' PutIntent (Maybe Text)

A description of the intent.

piName :: Lens' PutIntent Text

The name of the intent. The name is not case sensitive. The name can't match a built-in intent name, or a built-in intent name with "AMAZON." removed. For example, because there is a built-in intent called AMAZON.HelpIntent , you can't create a custom intent called HelpIntent . For a list of built-in intents, see Standard Built-in Intents in the Alexa Skills Kit .

Destructuring the Response

putIntentResponse

Arguments

:: Int
pirsResponseStatus

-> PutIntentResponse

Creates a value of PutIntentResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pirsFulfillmentActivity - If defined in the intent, Amazon Lex invokes this Lambda function to fulfill the intent after the user provides all of the information required by the intent.

pirsSlots - An array of intent slots that are configured for the intent.

pirsRejectionStatement - If the user answers "no" to the question defined in confirmationPrompt Amazon Lex responds with this statement to acknowledge that the intent was canceled.

pirsChecksum - Checksum of the > LATEST version of the intent created or updated.

pirsConclusionStatement - After the Lambda function specified in thefulfillmentActivity intent fulfills the intent, Amazon Lex conveys this statement to the user.

pirsSampleUtterances - An array of sample utterances that are configured for the intent.

pirsParentIntentSignature - A unique identifier for the built-in intent that this intent is based on.

pirsCreatedDate - The date that the intent was created.

pirsName - The name of the intent.

pirsVersion - The version of the intent. For a new intent, the version is always > LATEST .

pirsFollowUpPrompt - If defined in the intent, Amazon Lex uses this prompt to solicit additional user activity after the intent is fulfilled.

pirsLastUpdatedDate - The date that the intent was updated. When you create a resource, the creation date and last update dates are the same.

pirsConfirmationPrompt - If defined in the intent, Amazon Lex prompts the user to confirm the intent before fulfilling it.

pirsDialogCodeHook - If defined in the intent, Amazon Lex invokes this Lambda function for each user input.

pirsDescription - A description of the intent.

pirsResponseStatus - -- | The response status code.

data PutIntentResponse

See: putIntentResponse smart constructor.

Response Lenses

pirsFulfillmentActivity :: Lens' PutIntentResponse (Maybe FulfillmentActivity)

If defined in the intent, Amazon Lex invokes this Lambda function to fulfill the intent after the user provides all of the information required by the intent.

pirsSlots :: Lens' PutIntentResponse [Slot]

An array of intent slots that are configured for the intent.

pirsRejectionStatement :: Lens' PutIntentResponse (Maybe Statement)

If the user answers "no" to the question defined in confirmationPrompt Amazon Lex responds with this statement to acknowledge that the intent was canceled.

pirsChecksum :: Lens' PutIntentResponse (Maybe Text)

Checksum of the > LATEST version of the intent created or updated.

pirsConclusionStatement :: Lens' PutIntentResponse (Maybe Statement)

After the Lambda function specified in thefulfillmentActivity intent fulfills the intent, Amazon Lex conveys this statement to the user.

pirsSampleUtterances :: Lens' PutIntentResponse [Text]

An array of sample utterances that are configured for the intent.

pirsParentIntentSignature :: Lens' PutIntentResponse (Maybe Text)

A unique identifier for the built-in intent that this intent is based on.

pirsCreatedDate :: Lens' PutIntentResponse (Maybe UTCTime)

The date that the intent was created.

pirsName :: Lens' PutIntentResponse (Maybe Text)

The name of the intent.

pirsVersion :: Lens' PutIntentResponse (Maybe Text)

The version of the intent. For a new intent, the version is always > LATEST .

pirsFollowUpPrompt :: Lens' PutIntentResponse (Maybe FollowUpPrompt)

If defined in the intent, Amazon Lex uses this prompt to solicit additional user activity after the intent is fulfilled.

pirsLastUpdatedDate :: Lens' PutIntentResponse (Maybe UTCTime)

The date that the intent was updated. When you create a resource, the creation date and last update dates are the same.

pirsConfirmationPrompt :: Lens' PutIntentResponse (Maybe Prompt)

If defined in the intent, Amazon Lex prompts the user to confirm the intent before fulfilling it.

pirsDialogCodeHook :: Lens' PutIntentResponse (Maybe CodeHook)

If defined in the intent, Amazon Lex invokes this Lambda function for each user input.

pirsDescription :: Lens' PutIntentResponse (Maybe Text)

A description of the intent.

pirsResponseStatus :: Lens' PutIntentResponse Int

- | The response status code.