Network.AWS.LexModels.PutSlotType

Description

Creates a custom slot type or replaces an existing custom slot type.

To create a custom slot type, specify a name for the slot type and a set of enumeration values, which are the values that a slot of this type can assume. For more information, see 'how-it-works' .

If you specify the name of an existing slot type, the fields in the request replace the existing values in the > LATEST version of the slot type. Amazon Lex removes the fields that you don't provide in the request. If you don't specify required fields, Amazon Lex throws an exception.

This operation requires permissions for the lex:PutSlotType action.

Creating a Request

putSlotType

Arguments

:: Text
pstName

-> PutSlotType

Creates a value of PutSlotType with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pstChecksum - Identifies a specific revision of the > LATEST version. When you create a new slot type, leave the checksum field blank. If you specify a checksum you get a BadRequestException exception. When you want to update a slot type, set the checksum field to the checksum of the most recent revision of the > LATEST version. If you don't specify the checksum field, or if the checksum does not match the > LATEST version, you get a PreconditionFailedException exception.

pstDescription - A description of the slot type.

pstEnumerationValues - A list of EnumerationValue objects that defines the values that the slot type can take.

pstName - The name of the slot type. The name is not case sensitive. The name can't match a built-in slot type name, or a built-in slot type name with "AMAZON." removed. For example, because there is a built-in slot type called AMAZON.DATE , you can't create a custom slot type called DATE . For a list of built-in slot types, see Slot Type Reference in the Alexa Skills Kit .

data PutSlotType

See: putSlotType smart constructor.

Request Lenses

pstChecksum :: Lens' PutSlotType (Maybe Text)

Identifies a specific revision of the > LATEST version. When you create a new slot type, leave the checksum field blank. If you specify a checksum you get a BadRequestException exception. When you want to update a slot type, set the checksum field to the checksum of the most recent revision of the > LATEST version. If you don't specify the checksum field, or if the checksum does not match the > LATEST version, you get a PreconditionFailedException exception.

pstDescription :: Lens' PutSlotType (Maybe Text)

A description of the slot type.

pstEnumerationValues :: Lens' PutSlotType (Maybe (NonEmpty EnumerationValue))

A list of EnumerationValue objects that defines the values that the slot type can take.

pstName :: Lens' PutSlotType Text

The name of the slot type. The name is not case sensitive. The name can't match a built-in slot type name, or a built-in slot type name with "AMAZON." removed. For example, because there is a built-in slot type called AMAZON.DATE , you can't create a custom slot type called DATE . For a list of built-in slot types, see Slot Type Reference in the Alexa Skills Kit .

Destructuring the Response

putSlotTypeResponse

Arguments

:: Int
pstrsResponseStatus

-> PutSlotTypeResponse

Creates a value of PutSlotTypeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pstrsChecksum - Checksum of the > LATEST version of the slot type.

pstrsCreatedDate - The date that the slot type was created.

pstrsName - The name of the slot type.

pstrsVersion - The version of the slot type. For a new slot type, the version is always > LATEST .

pstrsLastUpdatedDate - The date that the slot type was updated. When you create a slot type, the creation date and last update date are the same.

pstrsDescription - A description of the slot type.

pstrsEnumerationValues - A list of EnumerationValue objects that defines the values that the slot type can take.

pstrsResponseStatus - -- | The response status code.

data PutSlotTypeResponse

See: putSlotTypeResponse smart constructor.

Response Lenses

pstrsChecksum :: Lens' PutSlotTypeResponse (Maybe Text)

Checksum of the > LATEST version of the slot type.

pstrsCreatedDate :: Lens' PutSlotTypeResponse (Maybe UTCTime)

The date that the slot type was created.

pstrsName :: Lens' PutSlotTypeResponse (Maybe Text)

The name of the slot type.

pstrsVersion :: Lens' PutSlotTypeResponse (Maybe Text)

The version of the slot type. For a new slot type, the version is always > LATEST .

pstrsLastUpdatedDate :: Lens' PutSlotTypeResponse (Maybe UTCTime)

The date that the slot type was updated. When you create a slot type, the creation date and last update date are the same.

pstrsDescription :: Lens' PutSlotTypeResponse (Maybe Text)

A description of the slot type.

pstrsEnumerationValues :: Lens' PutSlotTypeResponse (Maybe (NonEmpty EnumerationValue))

A list of EnumerationValue objects that defines the values that the slot type can take.

pstrsResponseStatus :: Lens' PutSlotTypeResponse Int

- | The response status code.